### 2022-06-14 #4478 @editorialbot
[PRE REVIEW]: Pycrostates: a Python library to study EEG microstates
**Submitting author:** @vferat (<a href="http://orcid.org/0000-0003-1952-7657">Victor Férat</a>)
**Repository:** https://github.com/vferat/pycrostates
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.1
