### 2022-06-15 #4478 @arfon
@vferat – please see the example paper for how you can denote corresponding and co-authors (this has recently been updated): https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
