+++ paper/paper.bib
@@ -54,6 +54,7 @@ arXiv:1309.0238 [cs]},
 
 @article{brunet_spatiotemporal_nodate,
 	title = {Spatiotemporal {Analysis} of {Multichannel} {EEG}: {CARTOOL}},
+	doi = {10.1155/2011/813870},
 	language = {en},
 	journal = {Computational Intelligence and Neuroscience},
 	author = {Brunet, Denis and Murray, Micah M and Michel, Christoph M},
@@ -117,7 +118,7 @@ pages = {577-593},
 year = {2018},
 note = {Brain Connectivity Dynamics},
 issn = {1053-8119},
-doi = {https://doi.org/10.1016/j.neuroimage.2017.11.062},
+doi = {10.1016/j.neuroimage.2017.11.062},
 url = {https://www.sciencedirect.com/science/article/pii/S105381191731008X},
 author = {Christoph M. Michel and Thomas Koenig},
 keywords = {EEG microstates, Resting state networks, Consciousness, Psychiatric disease, State-dependent information processing, Metastability},
@@ -134,6 +135,7 @@ abstract = {The present review discusses a well-established method for character
 @article{lehmann1971multichannel,
   title={Multichannel topography of human alpha EEG fields},
   author={Lehmann, D},
+  doi={10.1016/0013-4694(71)90165-9},
   journal={Electroencephalography and clinical neurophysiology},
   volume={31},
   number={5},
