### 2022-06-17 #4478 @vferat
> @vferat – please see the example paper for how you can denote corresponding and co-authors (this has recently been updated): https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography

@arfon Should be all good now
