### 2022-06-18 #4478 @arfon
:wave: @meg-simula – would you be able to handle this submission for JOSS?
