### 2022-06-25 eec473f
Bump py-actions/flake8 from 1 to 2 (#49)
Bumps [py-actions/flake8](https://github.com/py-actions/flake8) from 1 to 2.
- [Release notes](https://github.com/py-actions/flake8/releases)
- [Changelog](https://github.com/py-actions/flake8/blob/master/CHANGELOG.md)
- [Commits](https://github.com/py-actions/flake8/compare/v1...v2)

---
updated-dependencies:
- dependency-name: py-actions/flake8
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>

Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.
[…]
