+++ .github/workflows/code-style.yml
@@ -30,7 +30,7 @@ jobs:
       - name: Setup flake8 annotations
         uses: rbialon/flake8-annotations@v1
       - name: Run flake8
-        uses: py-actions/flake8@v1
+        uses: py-actions/flake8@v2
         with:
           path: "pycrostates"
       - name: Run isort
