### 2022-07-03 #4478 @Kevin-Mattheus-Moerman
@brainstorm, @rougier, @satra, @tuliofalmeida, @effigies, @szorowi1, @agricolab, @sappelhoff, @ajquinn, @vahid-sb, @cbrnr, @DominiqueMakowski, @mgbleichner would you be interested in reviewing this work featuring a Python library for EEG for [JOSS](https://joss.theoj.org/).
JOSS reviews focus on [the software](https://github.com/vferat/pycrostates) as well as [a short paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04478/joss.04478/10.21105.joss.04478.pdf). The review process takes place on GitHub. This here is a pre-review issue, where reviewers are assigned. Once we have enough reviewers I'll trigger a dedicated review issue, here is an example of an ongoing review: https://github.com/openjournals/joss-reviews/issues/4456#issuecomment-1170141621.

If interested please comment here, thanks!
