### 2022-07-03 #4478 @vahid-sb
@Kevin-Mattheus-Moerman Seems to be interesting and also in line with my interests in EEG analysis. I'd be happy to review it.
