### 2022-07-04 #4478 @sappelhoff
I have to decline for time reasons, but this reminded me of https://github.com/wmvanvliet/mne_microstates by @wmvanvliet who might want to review.
