### 2022-07-11 #4564 @editorialbot
[REVIEW]: Pycrostates: a Python library to study EEG microstates
**Submitting author:** @vferat (<a href="http://orcid.org/0000-0003-1952-7657">Victor Férat</a>)
**Repository:** https://github.com/vferat/pycrostates
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.1

@AJQuinn & @vahid-sb & @tuliofalmeida & @anshu-97, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
