### 2022-07-14 vferat/pycrostates#51 @tuliofalmeida
Review JOSS
Hi. This is an issue created as part of [pycrostates JOSS Review](https://github.com/openjournals/joss-reviews/issues/4564)

I've been doing some tests and following the examples you provided in the documentation. First, congratulations for the work, I liked the way you developed the package and I believe it will be very useful.

All the things I tried worked fine, so far so good. I have few things to suggest that are quite simple. I mostly used google colab to test the package and I missed your examples of adding the installation of the necessary packages.

I had to add the following lines in the examples:

```
!pip install -U mne
!pip install pycrostates
!pip install pymatreader

import pymatreader
```

The only example where I had to change some things to run was "plot_extract_gfp_peaks". In this I managed to solve the errors by performing the installation through your repository. Also the "dark mode" of the documentation some topics were hard to see (I don't know if it was a matter of my browser settings).
