### 2022-07-15 dae40c5
add note to the examples
