+++ tutorials/group_level_analysis/plot_group_level_with_individual_clusters.py
@@ -7,7 +7,16 @@ by computing group-level topographies based on individual clusters.
 """
 
 #%%
-# We first start by importing some EEG data
+# .. note::
+#
+#     The lemon datasets is composed of EEGLAB files. To use the MNE reader
+#     :func:`mne.io.read_raw_eeglab`, the ``pymatreader`` optional dependency
+#     is required. Use the following installation method appropriate for your
+#     environment:
+#     - ``pip install pymatreader```
+#     - ``conda install -c conda-forge pymatreader```
+#     Note that an environment created via the MNE installers includes
+#     ``pymatreader`` by default.
 
 from mne.io import read_raw_eeglab
 
