+++ tutorials/clustering/plot_modK.py
@@ -6,6 +6,18 @@ This tutorial introduces the :class:`pycrostates.clustering.ModKMeans`
 structure in detail.
 """
 
+#%%
+# .. note::
+#
+#     The lemon datasets is composed of EEGLAB files. To use the MNE reader
+#     :func:`mne.io.read_raw_eeglab`, the ``pymatreader`` optional dependency
+#     is required. Use the following installation method appropriate for your
+#     environment:
+#     - ``pip install pymatreader```
+#     - ``conda install -c conda-forge pymatreader```
+#     Note that an environment created via the MNE installers includes
+#     ``pymatreader`` by default.
+
 from mne.io import read_raw_eeglab
 
 from pycrostates.cluster import ModKMeans
