+++ tutorials/preprocessing/plot_extract_gfp_peaks.py
@@ -6,7 +6,16 @@ This example demonstrates how to extract global field power (gfp) peaks for an e
 """
 
 #%%
-# We start by loading some example data:
+# .. note::
+#
+#     The lemon datasets is composed of EEGLAB files. To use the MNE reader
+#     :func:`mne.io.read_raw_eeglab`, the ``pymatreader`` optional dependency
+#     is required. Use the following installation method appropriate for your
+#     environment:
+#     - ``pip install pymatreader```
+#     - ``conda install -c conda-forge pymatreader```
+#     Note that an environment created via the MNE installers includes
+#     ``pymatreader`` by default.
 
 import mne
 from mne.io import read_raw_eeglab
