### 2022-07-15 vferat/pycrostates#51 @mscheltienne
I added this note for the examples.

```
# .. note::
#
#     The lemon datasets is composed of EEGLAB files. To use the MNE reader
#     :func:`mne.io.read_raw_eeglab`, the ``pymatreader`` optional dependency
#     is required. Use the following installation method appropriate for your
#     environment:
#     - ``pip install pymatreader```
#     - ``conda install -c conda-forge pymatreader```
#     Note that an environment created via the MNE installers includes
#     ``pymatreader`` by default.
```

Also, if MNE is installed with the [standalone installers](https://mne.tools/stable/install/index.html), it will include all its optional dependencies and `pycrostates`, as of the next version 1.1.0 which should be released very shortly (https://github.com/mne-tools/mne-installers/pull/142).
