+++ pycrostates/datasets/lemon/lemon.py
@@ -54,8 +54,10 @@ def data_path(subject_id: str, condition: str) -> Path:
     :func:`mne.io.read_raw_eeglab`, the ``pymatreader`` optional dependency
     is required. Use the following installation method appropriate for your
     environment:
-    - ``pip install pymatreader```
-    - ``conda install -c conda-forge pymatreader```
+
+    - ``pip install pymatreader``
+    - ``conda install -c conda-forge pymatreader``
+
     Note that an environment created via the MNE installers includes
     ``pymatreader`` by default.
     """  # noqa: E501
