+++ tutorials/preprocessing/plot_extract_gfp_peaks.py
@@ -12,8 +12,10 @@ This example demonstrates how to extract global field power (gfp) peaks for an e
 #     :func:`mne.io.read_raw_eeglab`, the ``pymatreader`` optional dependency
 #     is required. Use the following installation method appropriate for your
 #     environment:
-#     - ``pip install pymatreader```
-#     - ``conda install -c conda-forge pymatreader```
+#
+#     - ``pip install pymatreader``
+#     - ``conda install -c conda-forge pymatreader``
+#
 #     Note that an environment created via the MNE installers includes
 #     ``pymatreader`` by default.
 
