+++ tutorials/metrics/plot_metrics.py
@@ -17,8 +17,10 @@ and the variance expressed by the decomposition.
 #     :func:`mne.io.read_raw_eeglab`, the ``pymatreader`` optional dependency
 #     is required. Use the following installation method appropriate for your
 #     environment:
-#     - ``pip install pymatreader```
-#     - ``conda install -c conda-forge pymatreader```
+#
+#     - ``pip install pymatreader``
+#     - ``conda install -c conda-forge pymatreader``
+#
 #     Note that an environment created via the MNE installers includes
 #     ``pymatreader`` by default.
 
