### 2022-07-15 44a5787
Add note for `mne.io.read_raw_eeglab` calls requiring `pymatreader` (#52)
* add note to the examples

* add note to API

* fix typos
