### 2022-07-21 84f52e9
edit the versionadded and versionchanged from docdict entries of MNE
