### 2022-07-26 vferat/pycrostates#51 @tuliofalmeida
Sorry for the delay! I agree with using the note, I think it will work fine and I didn't know about the pydata-sphinx-theme issue, sorry.

About the "plot_extract_gfp_peaks" after installing pycrostates and pymatreader I found this error:

`
from pycrostates.preprocessing import extract_gfp_peaks
gfp_data = extract_gfp_peaks(raw, min_peak_distance=3)
gfp_data

[extract_gfp_peaks.extract_gfp_peaks] INFO: 14713 GFP peaks extracted out of 119278 samples (12.34% of the original data).

TemplateNotFound                          Traceback (most recent call last)

[/usr/local/lib/python3.7/dist-packages/IPython/core/formatters.py](https://localhost:8080/#) in __call__(self, obj)
    336             method = get_real_method(obj, self.print_method)
    337             if method is not None:
--> 338                 return method()
    339             return None
    340         else:

4 frames

[/usr/local/lib/python3.7/dist-packages/jinja2/loaders.py](https://localhost:8080/#) in get_source(self, environment, template)
    247
    248         if not self.provider.has_resource(p):
--> 249             raise TemplateNotFound(template)
    250
    251         filename = uptodate = None

TemplateNotFound: ChData.html.jinja`

To proceed I cloned the github repository.
