### 2022-07-26 vferat/pycrostates#51 @mscheltienne
Good find, the `.jinja` files with the HTML templates for a couple of classes (among which `ChData`) are not packaged. I'll fix it.
