### 2022-07-26 a4f4511
[MNT] Package jinja templates (#54)
