### 2022-07-28 #4564 @vahid-sb
> Thanks @vahid-sb for the update. Note you may need to retry the checklist generation later, if it does not appear shortly.
>
> @openjournals/dev the checklist is not showing up yet here it seems.

No it has not. I'd try it again tomorrow.
