### 2022-08-09 377c4be
Add citations and contributing details
