### 2022-08-09 2d4300d
add white logo
