### 2022-08-10 659467c
fix link to install in readme
