### 2022-08-10 8b7f09a
improve doc build configuration
