### 2022-08-10 5e9c98b
update install page with a tab design
