### 2022-08-10 f551449
move links to a separate file
