### 2022-08-10 92cd567
add gh workflow to build the doc
