### 2022-08-10 2c610fa
revert min sphinx version
