+++ tutorials/group_level_analysis/plot_group_level_with_individual_gfp_peaks.py
@@ -1,6 +1,6 @@
 """
 Group level analysis from individual gfp peaks
-====================
+==============================================
 
 In this tutorial, we will learn how to conduct group level analysis
 by computing group level topogrpahies based on individual gfp peaks.
