### 2022-08-10 52621a8
ignore doc validation for method that matches dict subclasses
