### 2022-08-10 3770e7b
fix reference in clustering tutorial
