+++ tutorials/clustering/plot_modK.py
@@ -2,7 +2,7 @@
 The ModKmeans object
 ====================
 
-This tutorial introduces the :class:`pycrostates.clustering.ModKMeans`
+This tutorial introduces the :class:`pycrostates.cluster.ModKMeans`
 structure in detail.
 """
 
@@ -35,7 +35,7 @@ raw.set_eeg_reference('average')
 
 #%%
 # The modified Kmeans can be instantiated with the number of cluster centers
-# ``n_clusters`` to compute. By default, the modified Kmeans will only work
+# ``n_clusters`` to compute. By default, the modified K-means will only work
 # with EEG data, but this can be modified thanks to the ``picks`` parameter.
 # A random_state can be defined during class definition in order to have
 # reproducible results.
@@ -44,7 +44,7 @@ n_clusters = 5
 ModK = ModKMeans(n_clusters=n_clusters, random_state=42)
 
 #%%
-# Most methods need the modified Kmeans to be fitted. This can be done with
+# Most methods need the modified K-means to be fitted. This can be done with
 # either :class:`mne.io.Raw` or :class:`mne.epochs.Epochs` data structures.
 # Note that, depending on your setup, you can change ``n_jobs=1`` in order to
 # use parallel processing and reduce computation time.
@@ -53,9 +53,8 @@ ModK.fit(raw, n_jobs=5)
 
 #%%
 # Now that our algorithm is fitted, we can visualize the cluster centers, also
-# called microstate maps or microstate topographies using :meth:`ModK.plot`.
-# Note than this method uses the :class:`~mne.Info` object of the fitted
-# instance to display the topographies.
+# called microstate maps or microstate topographies using
+# :meth:`pycrostates.cluster.ModKMeans.plot`.
 
 ModK.plot()
 
@@ -66,19 +65,21 @@ ModK.plot()
 ModK.cluster_centers_
 
 #%%
-# Clusters centers can be reordered using :meth:`ModK.reorder_clusters`:
+# Clusters centers can be reordered using
+# :meth:`pycrostates.cluster.ModKMeans.reorder_clusters`.
 
 ModK.reorder_clusters(order=[3, 2, 4, 0, 1])
 ModK.plot()
 
 #%%
-# and renamed using :meth:`ModK.rename_clusters`:
+# and renamed using :meth:`pycrostates.cluster.ModKMeans.rename_clusters`.
 
 ModK.rename_clusters(new_names=['A', 'B', 'C', 'D', 'F'])
 ModK.plot()
 
 #%%
-# Maps polarities can be inverted thanks to :meth:`ModK.invert_polarity`
+# Maps polarities can be inverted using
+# :meth:`pycrostates.cluster.ModKMeans.invert_polarity`.
 # method. Note that it only affects visualization, it has not effect during
 # backfitting as polarities are ignored.
 
@@ -86,16 +87,16 @@ ModK.invert_polarity([False, False, True, True, False])
 ModK.plot()
 
 #%%
-# Finally, the modified Kmeans can be used to predict the microstates
-# segmentation using the :meth:`ModK.predict` method. By default, semgents
-# annotated as bad will not be labeled, but this behavior can be changed with
-# the ``reject_by_annotation`` argument. Smoothing can be performed on the
-# output sequence by setting the ``factor`` argument ``> 0`` (no smoothing by
-# default ``factor=0``) while the ``half_window_size`` parameter is used to
-# specify the smoothing temporal span. Finally, the ``reject_edges`` argument
-# allows not to assign the first and last segment of each record (or each
-# epoch) as these can be incomplete. It should have little impact for raw, but
-# can be important when working with epochs.
+# Finally, the modified K-means can be used to predict the microstates
+# segmentation using the :meth:`pycrostates.cluster.ModKMeans.predict` method.
[…]
