### 2022-08-10 50a8fbc
remove outdated clustering.rst (not included in any toctree)
