+++ docs/source/clustering.rst
@@ -1,8 +1,0 @@
-Clustering
-==========
-
-Pycrostates is meant to be scalable and aims at proposing different clustering
-algorithms adapted to the analysis of microstates.
-At the moment, the implemented algorithms are:
-
-*  :class:`pycrostates.cluster.ModKMeans`
