+++ tutorials/preprocessing/plot_resampling.py
@@ -2,7 +2,7 @@
 Resampling
 ==========
 
-This example demonstrates how to resemple a recording.
+This example demonstrates how to resample a recording.
 """
 
 #%%
@@ -33,23 +33,23 @@ raw.pick('eeg')
 raw.set_eeg_reference('average')
 
 #%%
-# We can now use the :func:`~pycrostates.preprocessing.resample`function to
-# draw n_resamples of n_samples for our recording where n_sample defines the
-# number of sample contained in each epoch and n_resamples defines the number of
-# epochs to draw.
+# We can now use the :func:`~pycrostates.preprocessing.resample` function to
+# draw ``n_resamples`` of ``n_samples`` for our recording where ``n_samples``
+# defines the number of sample contained in each epoch and ``n_resamples``
+# defines the number of epochs to draw.
 
 from pycrostates.preprocessing import resample
 resamples = resample(raw, n_resamples=10, n_samples=150, random_state=40)
 resamples
 
 #%%
-# We can also use the 'coverage' parameter to automatically compute one of the
-# two previous parameters based on the amount of original data we want to
-# cover. For example by setting n_resamples and coverage:
+# We can also use the ``coverage`` parameter to automatically compute one of
+# the two previous parameters based on the amount of original data we want to
+# cover. For example by setting ``n_resamples`` and ``coverage``:
 
 resamples = resample(raw, n_resamples=10, coverage=0.5, random_state=40)
 resamples
-#%% or by setting n_samples and coverage:
+#%% or by setting ``n_samples`` and ``coverage``:
 
 resamples = resample(raw, n_samples=150, coverage=0.5, random_state=40)
 resamples
