### 2022-08-10 7ff6005
more docstring fixes
