+++ docs/source/index.rst
@@ -1,7 +1,20 @@
+.. include:: ../links.inc
+
 Pycrostates
 ===========
 
-.. include:: overview.rst
+An open-source Python package for analyzing EEG microstates developed to work
+with `MNE-python <mne stable_>`_.
+
+.. image:: _static/img/FBMLAB_logo.png
+   :width: 200
+   :alt: FBMLAB logo
+   :target: https://www.unige.ch/medecine/neuf/en/researc/grecherche/christoph-michel
+
+.. image:: _static/img/UNIGE_logo.png
+   :width: 300
+   :alt: UNIGE logo
+   :target: https://www.unige.ch/en/university/presentation
 
 .. toctree::
    :hidden:
