### 2022-08-10 3f2aff0
update deprecated version specifier for readthedocs
