+++ .readthedocs.yml
@@ -14,9 +14,13 @@ formats: all
 
 # Optionally set the version of Python and requirements required to build your docs
 python:
-   version: 3.8
    install:
    - method: pip
      path: .
      extra_requirements:
        - docs
+
+build:
+  os: ubuntu-20.04
+  tools:
+    python: 3.8
