### 2022-08-10 90ed2a8
add fail on warning
