+++ .readthedocs.yml
@@ -24,3 +24,8 @@ build:
   os: ubuntu-20.04
   tools:
     python: 3.8
+
+sphinx:
+  builder: html
+  configuration: docs/source/conf.py
+  fail_on_warning: true
