+++ .readthedocs.yml
@@ -23,7 +23,7 @@ python:
 build:
   os: ubuntu-20.04
   tools:
-    python: 3.8
+    python: "3.8"
 
 sphinx:
   builder: html
