### 2022-08-10 d3c2120
more warning fixes: down to 30
