### 2022-08-10 ff0cbc5
up to 29 warnings but bibtex is working
