+++ pycrostates/datasets/lemon/lemon.py
@@ -17,7 +17,7 @@ from ...utils._config import get_config
 
 
 def data_path(subject_id: str, condition: str) -> Path:
-    """Get path to a local copy of preprocessed EEG recording from the LEMON dataset.
+    """Get path to a local copy of preprocessed EEG recording from the LEMON dataset\ :footcite:p:`babayan_mind-brain-body_2019`.
 
     Get path to a local copy of preprocessed EEG recording from the
     mind-brain-body dataset of MRI, EEG, cognition, emotion, and peripheral
@@ -56,10 +56,7 @@ def data_path(subject_id: str, condition: str) -> Path:
 
     References
     ----------
-    .. [1] Babayan, A., Erbey, M., Kumral, D. et al.
-           A mind-brain-body dataset of MRI, EEG, cognition, emotion,
-           and peripheral physiology in young and old adults. Sci Data 6, 180308 (2019).
-           https://doi.org/10.1038/sdata.2018.308
+    .. footbibliography::
     """  # noqa: E501
     _check_type(subject_id, (str,), "subject_id")
     _check_type(condition, (str,), "condition")
