+++ docs/references.bib
@@ -1,0 +1,62 @@
+% Encoding: UTF-8
+%
+% If available, include a DOI (preferred) *or* a URL for a given reference, but
+# not both, as the DOI turns into a link which is redundant with the URL.
+
+% MNE-C reference
+@article{GramfortEtAl2014,
+  title = {{{MNE}} Software for Processing {{MEG}} and {{EEG}} Data},
+  author = {Gramfort, Alexandre and Luessi, Martin and Larson, Eric and Engemann, Denis A. and Strohmeier, Daniel and Brodbeck, Christian and Parkkonen, Lauri and H{\"a}m{\"a}l{\"a}inen, Matti S.},
+  year = {2014},
+  volume = {86},
+  pages = {446--460},
+  doi = {10.1016/j.neuroimage.2013.10.027},
+  journal = {NeuroImage},
+}
+% MNE-Python reference
+@article{GramfortEtAl2013a,
+  title = {{{MEG}} and {{EEG}} Data Analysis with {{MNE}}-{{Python}}},
+  author = {Gramfort, Alexandre and Luessi, Martin and Larson, Eric and Engemann, Denis A. and Strohmeier, Daniel and Brodbeck, Christian and Goj, Roman and Jas, Mainak and Brooks, Teon and Parkkonen, Lauri and H{\"a}m{\"a}l{\"a}inen, Matti S.},
+  year = {2013},
+  volume = {7},
+  pages = {1--13},
+  doi = {10.3389/fnins.2013.00267},
+  journal = {Frontiers in Neuroscience},
+  number = {267}
+}
+% Michel 2018 reference
+@article{MICHEL2018577,
+  title = {EEG microstates as a tool for studying the temporal dynamics of whole-brain neuronal networks: A review},
+  journal = {NeuroImage},
+  volume = {180},
+  pages = {577-593},
+  year = {2018},
+  note = {Brain Connectivity Dynamics},
+  issn = {1053-8119},
+  doi = {10.1016/j.neuroimage.2017.11.062},
+  author = {Christoph M. Michel and Thomas Koenig},
+  keywords = {EEG microstates, Resting state networks, Consciousness, Psychiatric disease, State-dependent information processing, Metastability},
+}
+% Modified K-Means reference
+@article{391164,
+  author={Pascual-Marqui, R.D. and Michel, C.M. and Lehmann, D.},
+  journal={IEEE Transactions on Biomedical Engineering},
+  title={Segmentation of brain electrical activity into microstates: model estimation and validation},
+  year={1995},
+  volume={42},
+  number={7},
+  pages={658-665},
+  doi={10.1109/10.391164}
+}
+% Lemon dataset reference
+@article{babayan_mind-brain-body_2019,
+	title = {A mind-brain-body dataset of {MRI}, {EEG}, cognition, emotion, and peripheral physiology in young and old adults},
+	volume = {6},
+	issn = {2052-4463},
+	doi = {10.1038/sdata.2018.308},
+	number = {1},
+	journal = {Scientific Data},
+	author = {Babayan, Anahit and Erbey, Miray and Kumral, Deniz},
+	year = {2019},
+	pages = {180308},
+}
