### 2022-08-10 dd919ed
use r"" as a backslash is used to escape the space before the citation
