### 2022-08-11 5617961
remove config values which do nothing
