+++ docs/source/conf.py
@@ -168,8 +168,6 @@ numpydoc_validation_exclude = {  # regex to ignore during docstring check
 
 # -- sphinxcontrib-bibtex ----------------------------------------------------
 bibtex_bibfiles = ['../references.bib']
-bibtex_style = 'unsrt'
-bibtex_footbibliography_header = ''
 
 # -- sphinx-gallery ----------------------------------------------------------
 sphinx_gallery_conf = {
