### 2022-08-12 #4564 @AJQuinn
Hi @Kevin-Mattheus-Moerman and @vferat.

First, apologies for the horrendous delay in this review. I was not able to complete it prioir to taking time away from work for house and job moves... I hope this wasn't too frustrating.

I have completed a review and will add the details on the project repo. The short summary is that I'm very impressed with the code quality, rigour and testing used in this project. It seems robust and straightforward to use. On the other hand, the docstrings and tutorials on the website are very brief. I think this substantially increases the difficultly for new users and could be easily improved with a bit more context and guidance in the tutorials.

Many thanks
