### 2022-08-12 vferat/pycrostates#60 @AJQuinn
JOSS Review
Hello - apologies for the long delay in returning this review. I wasn't able to complete it prior to a very busy personal period. I hope this wasn't too frustrating....

I am very impressed with the package, package layout and very thorough testing. Bravo.

I have two comments for discussion/action. One big and one small...

Firstly, the documentation seems rushed, in contrast to the quality of the code. It is easy to leave this to the last minute but is really critical. I think this is particularly the case for a microstates toolbox as so many potential users are likely to be much more familiar with matlab-based pipelines. It is currently a bit short of this sentence in the paper "Finally, new users will benefits from the exhaustive documentation and tutorials describing the microstate analysis modules."

If we take the resampling tutorial as an example. This does demonstrate the main function and a bit about its API but doesn't explain when or why we would use resampling. I'm fairly famiiliar with microstate pipelines but I'm still not sure what the motivation for resampling is based on the tutorial. Can you show...
- Why a resampling function would be useful
- How can we decide on the critical parameters
- How can we tell if it has gone wrong
- Is there further reading that you'd recommend

It would be great to take another pass through the tutorials with these points in mind, I think it will make onboarding much easier for newer users. Currently, I think a non-expert in either python or microstates could get lost.

Similarly, it would be good to expand the docstrings of the most prominant functions and classes. Could possibly also use the 'Notes' section to give a bit more context about the purpose and context of the code.

Secondly, I think the numpy=>1.20 requirement is slightly off.  I had 1.20 installed and got an error that `numpy.typing.NDArray` couldn't be imported. Looks like this was introduced in 1.21 https://numpy.org/devdocs/reference/typing.html#numpy.typing.NDArray . I think setting this to numpy=>1.21 would be safe.

Great work overall - congratulations.
