### 2022-08-12 #4564 @Kevin-Mattheus-Moerman
@AJQuinn thanks for getting back to us, and for starting the review. Note you may have to call `@editorialbot generate my checklist` again, sometimes our bot is stuck on something else and it appears not to work. @openjournals/dev might be working on a fix for this but for the moment please try again.
