### 2022-08-14 4e374c4
Small typo fixes (#59)
* Small typo fixes

* Small typos in paper
