+++ pycrostates/io/ch_data.py
@@ -1,4 +1,4 @@
-"""Clas to handle no temporal but spatial data."""
+"""Class to handle no temporal but spatial data."""
 from copy import copy, deepcopy
 from typing import Any, Union
 
