### 2022-08-14 #4564 @vahid-sb
@Kevin-Mattheus-Moerman I just filled in the checklist. From my point of view, the software has passed all the requirements to be published in JOSS. There were a couple of minor issues that the authors mentioned they'd resolve as part of another issue raised by another reviewer. My understanding is that my part of the review is completed, but if I have missed anything, please let me know.
