### 2022-08-14 307ffc4
Community friendly rework (#44)
* Add badges

skip ci

* Add contributing

* update Zenodo Badge

* Add citations and contributing details

* Fix style

* Update CONTRIBUTING.md

* Update README.md

Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>

* Update README.md

* minor changes

Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>
