### 2022-08-15 vferat/pycrostates#61 @anshu-97
JOSS Review
Hi there!
Apologies for the delay in my review and I do hope this hasn’t brought about too much inconvenience. Firstly, congratulations on launching the software! Overall, I was impressed by the ease I had in running all the examples and tutorials.
I will add some specific feedback I have, on how mostly the documentation and paper can be improved, directly as opened issues.
