### 2022-08-15 vferat/pycrostates#64 @anshu-97
[JOSS Review] Documentation
As part of #61

As mentioned by @AJQuinn in #60, as compared to how thorough the testing of the code itself appears to have been, the documentation seems relatively rushed and vague with little elaboration on the purpose and use of certain functions such as `resampling`, `ModK.invert_polarity` and `ModK.reorder_clusters`. Defaults set for certain important parameters such as the threshold used to locate for peaks in the `extract_gfp_peaks` function is also not mentioned in the API method documentation. While some of these default parameters are not set within the pycrostates function themselves (in this case, default threshold for peaks is set in `scipy.signal.find_peaks`), I believe briefly mentioning the set defaults in the docstrings or in Notes for some of the parameters important for preprocessing would be greatly appreciated by users, especially with regard to potential users that are unfamiliar with microstates or EEG analysis.

Furthermore, more guidance can also be provided for new users of microstates analysis with regards to the different clustering methods and evaluation metrics available for analysing cluster quality and choosing the optimal number of clusters. For instance, in the tutorial for choosing the number of cluster centers, it is briefly mentioned that “_compromises have to be made when considering the type of metric to use when evaluating cluster quality_”. However, there is no further elaboration given on what these compromises might be, nor are there any explanations or examples provided on what the different metrics purport to measure as an indicator of cluster quality. Rather than simply providing examples to show that the different metrics offer different clustering solutions for the same set of data, it would perhaps be more effective to contextualize and describe what the various evaluation metrics are measuring as indicators of cluster quality before providing the example of comparison between the different methods.
