### 2022-08-15 vferat/pycrostates#65 @anshu-97
[JOSS Review] Software Paper
As part of #61

Overall, the paper gives a concise and clear description of what microstates are and what microstate analysis entails. However, it is not as clear what microstate analysis is useful for and how it is currently being implemented in research. By extension, the need for a software that wholly serves a seemingly niche form of spatiotemporal analysis is ambiguous. That being the case, providing elaboration on the purpose of microstate analysis such as by highlighting its potential applications for real-world scenarios, its use over other forms of EEG analysis, the current gaps in the field that needs to be addressed and how this software seeks to do so etc., would more prominently showcase the need for the software more clearly and convincingly, especially for a non-specialist audience. For instance, in the second last line of the paragraph under _Statement of Need_, it is stated that Pycrostates _“is able to replicate most of the analyses proposed in the literature (Michel et al., 2018)_.”. Does this statement then imply that replicability is an issue facing the field that Pycrostates seeks to address? Are the other available packages/modules not able to replicate such past analyses and if so, what are the features Pycrostate provides that allow it to do so? These are some questions that could perhaps be pondered over to clearly point out to readers and potential users on the purpose, efficacy and usability of the software.

Moreover, the justification of the software’s usability over other microstate analyses software (i.e., that the MNE-microstates package has “_a limited number of features and is not maintained on a regular basis_”) could appear as a quick dismissal of a usable software. While there is a very brief overview on all the modules and sub-modules offered in the Pycrostates package, it is not clear how the utility of the specific features available in Pycrostates extend beyond that found in other packages. Given that the limited features of other packages serves as one of the two main reasons that justify the need for this software, the functionality and utility of its additional features should be very explicitly fleshed out for the justification to hold. For instance, providing comparisons between the generated output (e.g., are the microstate map plots clearer and/or more detailed) or in the way in which the algorithm is executed (e.g., is the code written more efficiently such that computation time is lesser etc.,) in the form of example codes in the paper could aid in highlighting the distinctions between Pycrostates and other available microstate analysis modules and packages. Though a software comparison table (e.g., [here]( https://www.sciencedirect.com/science/article/pii/S0169260720315510?via%3Dihub#tbl0003)) or [benchmark graphs]( https://sleepecg.readthedocs.io/en/stable/heartbeat_detection.html#benchmarks) would be great, any improvement of the presentation of Pycrostates within its software environment would be welcome.

Additionally, beyond the mention of mne-microstates and Wegner’s Python2 implementation of microstate analysis, the theoretical comparison with other Python packages implementing microstates have not been discussed, such as [KeyPy]( https://github.com/keyinst/keypy) (although keypy is no longer being maintained, I believe it is still worthy of a mention since it implements similar features for microstate analysis) and [NeuroKit2](https://github.com/neuropsychology/NeuroKit) (disclaimer: I am a maintainer of NK 😊).
