### 2022-08-15 vferat/pycrostates#60 @mscheltienne
Thank you very much for your review!

As you saw, the priority was put on the code and on the functionalities. All the items we worked on for the 0.1.0 release are code-related, c.f. https://github.com/vferat/pycrostates/issues/20

> It would be great to take another pass through the tutorials with these points in mind, I think it will make onboarding much easier for newer users. Currently, I think a non-expert in either python or microstates could get lost.

Agree! We will improve the documentation, little by little, but as this is a very time-consuming task, this item can not be fixed right away.

------------------

Thanks for the `numpy` minimum requirements, I'll fix it.
