### 2022-08-16 40837fa
add bool to x-ref to resolve to class:bool and not to std:label which renders bool x-ref correctly
