### 2022-08-20 bc26faa
Fix default picks
