### 2022-08-20 vferat/pycrostates#51 @mscheltienne
FYI, I added the `style.css` of MNE for `pydata-sphinx-theme` as part of #58, Daniel McCloy did a great job tuning the style for both the light and the dark theme.
