### 2022-08-20 7041108
remove non-needed arguments
