### 2022-08-21 d203ac8
Add _check_picks [WIP]
