### 2022-08-21 3d1a403
Update fit picks
