+++ pycrostates/cluster/_base.py
@@ -170,12 +170,13 @@ class _BaseCluster(ABC, ChannelsMixin, ContainsMixin, MontageMixin):
 
     def _check_picks(self, inst, picks):
         inst_ = inst.copy()
+        inst_.pick(picks)
         if len(inst_.get_channel_types(unique=True)) != 1:
             ch_types = inst_.get_channel_types(unique=False)
             ch_types, counts =  np.unique(ch_types, return_counts=True)
-            channels_msg = ', '.join('%s %s channels' % t for t in zip(counts, ch_types))
+            channels_msg = ', '.join('%s %s channel(s)' % t for t in zip(counts, ch_types))
             msg = f"Only one datatype can be fitted, but picks={picks} results in "
-                  + channels_msg
+            msg += channels_msg
             raise ValueError(msg)
 
     @abstractmethod
