+++ pycrostates/io/ch_data.py
@@ -115,7 +115,7 @@ class ChData(CHData, ChannelsMixin, ContainsMixin, MontageMixin):
         data = self._data.copy()
         return data[picks, :]
 
-    def pick(self, picks, none="all", exclude="bads"):
+    def pick(self, picks, exclude="bads"):
         """Pick a subset of channels.
         Parameters
         ----------
@@ -131,7 +131,7 @@ class ChData(CHData, ChannelsMixin, ContainsMixin, MontageMixin):
             The modified instance.
         """
         from .meas_info import ChInfo
-        picks = _picks_to_idx(self._info, picks, none=none, exclude=exclude)
+        picks = _picks_to_idx(self._info, picks, exclude=exclude)
         data = self._data.copy()
         data = data[picks, :]
 
@@ -139,6 +139,7 @@ class ChData(CHData, ChannelsMixin, ContainsMixin, MontageMixin):
 
         self._data = data
         self._info = info if isinstance(info, ChInfo) else ChInfo(info)
+        return(self)
 
     # --------------------------------------------------------------------
     @property
