### 2022-08-21 127eed0
Add test_fit_picks
