+++ pycrostates/cluster/tests/test_kmeans.py
@@ -798,6 +798,7 @@ def test_fit_with_bads(caplog):
     assert "Channels EEG 001, EEG 002 are set as bads" in caplog.text
     caplog.clear()
 
+
 def test_fit_picks():
     raw = raw_meg.copy().pick_types(meg=True, eeg=True, eog=True)
     ModK_ = ModKMeans(
@@ -815,6 +816,7 @@ def test_fit_picks():
         ModK_.fit(raw, picks="data")  # fails -> eeg + grad + mag
     ModK_.fit(raw, picks="eeg")  # works
     ModK_.fit(raw, picks="mag")  # works
+
 
 def test_predict(caplog):
     """Test predict method default behaviors."""
