+++ pycrostates/io/ch_data.py
@@ -131,6 +131,7 @@ class ChData(CHData, ChannelsMixin, ContainsMixin, MontageMixin):
             The modified instance.
         """
         from .meas_info import ChInfo
+
         picks = _picks_to_idx(self._info, picks, exclude=exclude)
         data = self._data.copy()
         data = data[picks, :]
@@ -139,7 +140,7 @@ class ChData(CHData, ChannelsMixin, ContainsMixin, MontageMixin):
 
         self._data = data
         self._info = info if isinstance(info, ChInfo) else ChInfo(info)
-        return(self)
+        return self
 
     # --------------------------------------------------------------------
     @property
