### 2022-08-22 e6b4dee
Update pycrostates/cluster/_base.py
Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>
