### 2022-08-22 00b94c2
Update pycrostates/cluster/_base.py
Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>
