### 2022-08-22 cd29641
Update pycrostates/io/ch_data.py
Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>
