### 2022-08-22 707ba43
better structure
