### 2022-08-22 0a1f99d
better autodoc settings
