### 2022-08-22 1f3ea91
fix and move staticmethod at the end of the file
