+++ pycrostates/cluster/tests/test_kmeans.py
@@ -49,7 +49,7 @@ n_clusters = 4
 ModK = ModKMeans(
     n_clusters=n_clusters, n_init=10, max_iter=100, tol=1e-4, random_state=1
 )
-ModK.fit(raw_eeg, n_jobs=1)
+ModK.fit(raw_eeg, picks="eeg", n_jobs=1)
 
 
 # pylint: disable=protected-access
@@ -800,6 +800,7 @@ def test_fit_with_bads(caplog):
 
 
 def test_fit_picks():
+    """Test fitting with different picks."""
     raw = raw_meg.copy().pick_types(meg=True, eeg=True, eog=True)
     ModK_ = ModKMeans(
         n_clusters=n_clusters,
