### 2022-08-22 06bc509
better error message
