+++ pycrostates/cluster/tests/test_kmeans.py
@@ -49,7 +49,7 @@ n_clusters = 4
 ModK = ModKMeans(
     n_clusters=n_clusters, n_init=10, max_iter=100, tol=1e-4, random_state=1
 )
-ModK.fit(raw_eeg, picks="eeg", n_jobs=1)
+ModK.fit(raw_eeg, n_jobs=1)
 
 
 # pylint: disable=protected-access
