### 2022-08-22 748df69
fix bibliographic warnings
