### 2022-08-22 3efff0b
Check channel for prediction
