### 2022-08-22 ebd5d1f
Use same picks for info and data in fit
