### 2022-08-22 a6f6b20
use predict bads yhen needed
