+++ pycrostates/cluster/_base.py
@@ -213,7 +213,6 @@ class _BaseCluster(ABC, ChannelsMixin, ContainsMixin, MontageMixin):
 
         # retrieve info
         info = inst.info
-
         # picks
         picks_bads_inc = _picks_to_idx(info, picks, none="all", exclude=[])
         picks = _picks_to_idx(info, picks, none="all", exclude="bads")
@@ -238,20 +237,36 @@ class _BaseCluster(ABC, ChannelsMixin, ContainsMixin, MontageMixin):
             )
             del msg
 
+        inst = inst.pick(picks)
+        if inst.info['bads'] != []:
+            if len(inst.info['bads']) == 1:
+                msg = (
+                    "Channel %s is bad but was explicity set in 'picks' "
+                    + "and therefore will be used during fitting."
+                )
+            else:
+                msg = (
+                    "Channels %s are bad but were explicity set in 'picks' "
+                    + "and therefore will be used during fitting."
+                )
+            logger.warning(
+                msg, ", ".join(inst.info['bads'])
+            )
+            del msg   
         # retrieve numpy array
         kwargs = (
             dict() if isinstance(inst, ChData) else dict(tmin=tmin, tmax=tmax)
         )
         if isinstance(inst, BaseRaw):
             kwargs["reject_by_annotation"] = reject_by_annotation
-        data = inst.get_data(picks=picks, **kwargs)
+        data = inst.get_data(**kwargs)
         # reshape if inst is Epochs
         if isinstance(inst, BaseEpochs):
             data = np.swapaxes(data, 0, 1)
             data = data.reshape(data.shape[0], -1)
 
         # store picks and info
-        self._info = ChInfo(info=pick_info(info, picks))
+        self._info = ChInfo(inst.info)
         self._fitted_data = data
         return data
 
