### 2022-08-22 a784400
fix autodoc entry in autosummary section
