### 2022-08-22 1c9ee74
Fix predict and fit with picks and bads
