+++ pycrostates/cluster/tests/test_kmeans.py
@@ -760,11 +760,11 @@ def test_fit_with_bads(caplog):
     ModK_.fit(raw_, n_jobs=1)
     _check_fitted(ModK_)
     _check_fitted_data_raw(ModK_._fitted_data, raw_, "eeg", None, None, "omit")
-    assert len(ModK_._info["bads"]) == 1
+    assert len(ModK_._info["bads"]) == 0
     assert "Channel EEG 001 is set as bad and ignored." in caplog.text
     caplog.clear()
 
-    # more than 1 bads
+    # test exclude bads
     raw_ = raw_eeg.copy()
     raw_.info["bads"] = [raw_.ch_names[0], raw_.ch_names[1]]
     ModK_ = ModKMeans(
@@ -774,10 +774,49 @@ def test_fit_with_bads(caplog):
         tol=1e-4,
         random_state=1,
     )
-    ModK_.fit(raw_, n_jobs=1)
+    ModK_.fit(raw_, picks="eeg", n_jobs=1)
     _check_fitted(ModK_)
     _check_fitted_data_raw(ModK_._fitted_data, raw_, "eeg", None, None, "omit")
+    assert len(ModK_._info["bads"]) == 0
+    assert "Channels EEG 001, EEG 002 are set as bads" in caplog.text
+    caplog.clear()
+
+    # test include bads
+    raw_ = raw_eeg.copy()
+    raw_.info["bads"] = [raw_.ch_names[0], raw_.ch_names[1]]
+    ModK_ = ModKMeans(
+        n_clusters=n_clusters,
+        n_init=10,
+        max_iter=100,
+        tol=1e-4,
+        random_state=1,
+    )
+    ModK_.fit(raw_, picks=raw_.ch_names, n_jobs=1)
+    _check_fitted(ModK_)
+    _check_fitted_data_raw(
+        ModK_._fitted_data,
+        raw_,
+        raw_.ch_names,
+        None,
+        None,
+        "omit",
+    )
     assert len(ModK_._info["bads"]) == 2
+
+    # Test on epochs
+    epochs_ = epochs_eeg.copy()
+    epochs_.info["bads"] = [epochs_.ch_names[0], epochs_.ch_names[1]]
+    ModK_ = ModKMeans(
+        n_clusters=n_clusters,
+        n_init=10,
+        max_iter=100,
+        tol=1e-4,
+        random_state=1,
+    )
+    ModK_.fit(epochs_, picks="eeg", n_jobs=1)
+    _check_fitted(ModK_)
+    _check_fitted_data_epochs(ModK_._fitted_data, epochs_, "eeg", None, None)
+    assert len(ModK_._info["bads"]) == 0
     assert "Channels EEG 001, EEG 002 are set as bads" in caplog.text
     caplog.clear()
 
@@ -791,12 +830,16 @@ def test_fit_with_bads(caplog):
         tol=1e-4,
         random_state=1,
     )
-    ModK_.fit(epochs_, n_jobs=1)
+    ModK_.fit(epochs_, picks=epochs_.ch_names, n_jobs=1)
     _check_fitted(ModK_)
-    _check_fitted_data_epochs(ModK_._fitted_data, epochs_, "eeg", None, None)
+    _check_fitted_data_epochs(
+        ModK_._fitted_data,
+        epochs_,
+        epochs_.ch_names,
+        None,
+        None,
+    )
     assert len(ModK_._info["bads"]) == 2
-    assert "Channels EEG 001, EEG 002 are set as bads" in caplog.text
-    caplog.clear()
 
 
 def test_fit_picks():
@@ -816,7 +859,7 @@ def test_fit_picks():
     with pytest.raises(ValueError, match="Only one datatype can be fitted"):
[…]
