### 2022-08-22 b1d87cc
Update test_fiff.py
