### 2022-08-23 6e99897
Delete (some) outdated tests
