+++ pycrostates/cluster/tests/test_kmeans.py
@@ -728,98 +728,7 @@ def test_fit_data_shapes():
     assert np.isclose(fitted_data_5_end, ModK_rej_5_end._fitted_data).all()
 
 
-def test_fit_with_bads(caplog):
-    """Test log messages emitted when fitting with bad channels."""
-    # 0 bads
-    raw_ = raw_eeg.copy()
-    raw_.info["bads"] = []
-    ModK_ = ModKMeans(
-        n_clusters=n_clusters,
-        n_init=10,
-        max_iter=100,
-        tol=1e-4,
-        random_state=1,
-    )
-    ModK_.fit(raw_, n_jobs=1)
-    _check_fitted(ModK_)
-    _check_fitted_data_raw(ModK_._fitted_data, raw_, "eeg", None, None, "omit")
-    assert len(ModK_._info["bads"]) == 0
-    assert "set as bad" not in caplog.text
-    caplog.clear()
-
-    # 1 bads
-    raw_ = raw_eeg.copy()
-    raw_.info["bads"] = [raw_.ch_names[0]]
-    ModK_ = ModKMeans(
-        n_clusters=n_clusters,
-        n_init=10,
-        max_iter=100,
-        tol=1e-4,
-        random_state=1,
-    )
-    ModK_.fit(raw_, n_jobs=1)
-    _check_fitted(ModK_)
-    _check_fitted_data_raw(ModK_._fitted_data, raw_, "eeg", None, None, "omit")
-    assert len(ModK_._info["bads"]) == 1
-    assert "Channel EEG 001 is set as bad and ignored." in caplog.text
-    caplog.clear()
-
-    # more than 1 bads
-    raw_ = raw_eeg.copy()
-    raw_.info["bads"] = [raw_.ch_names[0], raw_.ch_names[1]]
-    ModK_ = ModKMeans(
-        n_clusters=n_clusters,
-        n_init=10,
-        max_iter=100,
-        tol=1e-4,
-        random_state=1,
-    )
-    ModK_.fit(raw_, n_jobs=1)
-    _check_fitted(ModK_)
-    _check_fitted_data_raw(ModK_._fitted_data, raw_, "eeg", None, None, "omit")
-    assert len(ModK_._info["bads"]) == 2
-    assert "Channels EEG 001, EEG 002 are set as bads" in caplog.text
-    caplog.clear()
-
-    # Test on epochs
-    epochs_ = epochs_eeg.copy()
-    epochs_.info["bads"] = [epochs_.ch_names[0], epochs_.ch_names[1]]
-    ModK_ = ModKMeans(
-        n_clusters=n_clusters,
-        n_init=10,
-        max_iter=100,
-        tol=1e-4,
-        random_state=1,
-    )
-    ModK_.fit(epochs_, n_jobs=1)
-    _check_fitted(ModK_)
-    _check_fitted_data_epochs(ModK_._fitted_data, epochs_, "eeg", None, None)
-    assert len(ModK_._info["bads"]) == 2
-    assert "Channels EEG 001, EEG 002 are set as bads" in caplog.text
-    caplog.clear()
-
-
-def test_fit_picks():
-    """Test fitting with different picks."""
-    raw = raw_meg.copy().pick_types(meg=True, eeg=True, eog=True)
-    ModK_ = ModKMeans(
-        n_clusters=n_clusters,
-        n_init=10,
-        max_iter=100,
-        tol=1e-4,
-        random_state=1,
-    )
-    with pytest.raises(ValueError, match="Only one datatype can be fitted"):
-        ModK_.fit(raw, picks=None)  # fails -> MEG + EEG
-    with pytest.raises(ValueError, match="Only one datatype can be fitted"):
[…]
