### 2022-08-23 a2ec584
Add warning for picking bad channel during fitting
