### 2022-08-23 8c3476e
Add warning if bads channels in current fit in predict
