### 2022-08-23 a758e82
Handle missing and unused channels for prediction [WIP]
