+++ pycrostates/cluster/tests/test_kmeans.py
@@ -900,7 +900,9 @@ def test_picks_fit(caplog):
     )  # -> works, same channels as ModK.info
 
     raw_predict.info["bads"] = ["Fp1"]
-    with pytest.raises(ValueError, match="Missing channel Fp1"):
+    with pytest.raises(
+        ValueError, match="Cannot predict because channel Fp1 is not in"
+    ):
         ModK_.predict(raw_predict, picks="eeg")
     # fails, because ModK.info includes Fp1
     # which is bad in prediction instance
@@ -908,10 +910,15 @@ def test_picks_fit(caplog):
 
     ModK_.predict(raw_predict, picks=["Fp1", "CP1", "CP2"])
     # works, with warning because you predict based on a bad channel Fp1
-    assert "Fp1 is set as bad" in caplog.text
+    assert (
+        "Channel Fp1 is set as bad and will be used for prediction"
+        in caplog.text
+    )
     caplog.clear()
 
-    with pytest.raises(ValueError, match="Missing channel Fp1"):
+    with pytest.raises(
+        ValueError, match="Cannot predict because channel Fp1 is not in"
+    ):
         ModK_.predict(raw_predict, picks=["CP2", "CP1"])
     # fails, because ModK.info includes Fp1 which is missing
     # from prediction instance selection
@@ -927,7 +934,9 @@ def test_picks_fit(caplog):
     ModK_.predict(raw_predict, picks="eeg")
     # works, with warning because Fpz, Fp2 are missing from ModK.info
 
-    with pytest.raises(ValueError, match="Missing channel Fp1"):
+    with pytest.raises(
+        ValueError, match="Cannot predict because channel Fp1 is not in"
+    ):
         ModK_.predict(raw_predict, picks=["Fp2", "Fpz", "CP2", "CP1"])
     # -> fails, because ModK.info includes Fp1 which is missing
     # from prediction instance selection
@@ -935,11 +944,13 @@ def test_picks_fit(caplog):
 
     ModK_.predict(raw_predict, picks=["Fp1", "Fpz", "CP2", "CP1"])
     # -> works, with warning because Fpz is missing from ModK.info
-    assert (
-        "Picked channel Fp1 will not be used for prediction" not in caplog.text
-    )
-    caplog.clear()
 
+    raw_predict.drop_channels(["Fp1"])
+    with pytest.raises(ValueError, match="Missing channel Fp1"):
+        ModK_.predict(raw_predict, picks="eeg")
+    # -> fails, because Fp1 is used in ModK.info
+
+    # -> works, with warning because Fpz is missing from ModK.info
     # ---- Use bad channel Fp2 during fitting ----
     # Fit
     info = info_.copy()
@@ -973,14 +984,21 @@ def test_picks_fit(caplog):
     assert predict_warning in caplog.text
     caplog.clear()
 
-    with pytest.raises(ValueError, match="Missing channel Fp2"):
+    with pytest.raises(
+        ValueError, match="Cannot predict because channel Fp2 is not in"
+    ):
         ModK_.predict(raw_predict, picks=["Fp1", "CP2", "CP1"])
     # -> fails, because Fp2 is used in ModK.info
+    assert predict_warning in caplog.text
+    caplog.clear()
 
     raw_predict.info["bads"] = ["Fp2"]
-    with pytest.raises(ValueError, match="Missing channel Fp2"):
+    with pytest.raises(
+        ValueError, match="Cannot predict because channel Fp2 is not in"
+    ):
         ModK_.predict(raw_predict, picks="eeg")
     # -> fails, because Fp2 is used in ModK.info
+    assert predict_warning in caplog.text
     caplog.clear()
 
[…]
