### 2022-08-24 e943d03
update logo display
