### 2022-08-24 49f2bf8
Revert "TO DELETE  change image link"
This reverts commit c03c82256f0dfc2d1aea3583ad81a7c28cc97c51.
