### 2022-08-24 eb785ae
Logo proposal (#56)
* add logo to readme

* add white logo

* add logo to doc

* update logo display

* TO DELETE  change image link

* Revert "TO DELETE  change image link"

This reverts commit c03c82256f0dfc2d1aea3583ad81a7c28cc97c51.

* ad logo to website index

* move logo

* Remove title

Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>
Co-authored-by: Victor Férat <victor.ferat@live.Fr>
