### 2022-08-24 eef67e7
docs: add LouPlanchamp as a contributor for design (#68)
* docs: update README.md [skip ci]

* docs: update .all-contributorsrc [skip ci]

Co-authored-by: allcontributors[bot] <46447321+allcontributors[bot]@users.noreply.github.com>
