### 2022-09-02 ba8959a
better error message and rename _check_picks to _check_picks_uniqueness
