+++ pycrostates/cluster/_base.py
@@ -215,23 +215,25 @@ class _BaseCluster(ABC, ChannelsMixin, ContainsMixin, MontageMixin):
         info = inst.info
 
         # picks
-        picks_bads_inc = _picks_to_idx(info, picks, none="all", exclude=[])
-        picks = _picks_to_idx(info, picks, none="all", exclude="bads")
-        _BaseCluster._check_picks(inst.info, picks)  # picks has to be idx
+        picks_bads_inc = _picks_to_idx(
+            inst.info, picks, none="all", exclude=[]
+        )
+        picks = _picks_to_idx(inst.info, picks, none="all", exclude="bads")
+        _BaseCluster._check_picks_uniqueness(inst.info, picks)
         ch_not_used = set(picks_bads_inc) - set(picks)
         if len(ch_not_used) != 0:
             if len(ch_not_used) == 1:
                 msg = (
                     "Channel %s is set as bad and ignored. To include "
-                    + "it, either remove it from 'inst.info['bads'] or "
-                    + "provide its name explicitly in the 'picks' argument."
+                    "it, either remove it from 'inst.info['bads'] or "
+                    "provide it explicitly in the 'picks' argument."
                 )
             else:
                 msg = (
                     "Channels %s are set as bads and ignored. To "
-                    + "include them, either remove them from "
-                    + "'inst.info['bads'] or provide their names "
-                    + "explicitly in the 'picks' argument."
+                    "include them, either remove them from "
+                    "'inst.info['bads'] or provide them "
+                    "explicitly in the 'picks' argument."
                 )
             logger.warning(
                 msg, ", ".join(info["ch_names"][k] for k in ch_not_used)
@@ -1137,14 +1139,16 @@ class _BaseCluster(ABC, ChannelsMixin, ContainsMixin, MontageMixin):
         return n_clusters
 
     @staticmethod
-    def _check_picks(info, picks):
+    def _check_picks_uniqueness(info, picks):
         info = pick_info(info, picks, copy=True)
         if len(info.get_channel_types(unique=True)) != 1:
             ch_types = info.get_channel_types(unique=False)
             ch_types, counts = np.unique(ch_types, return_counts=True)
             channels_msg = ", ".join(
-                "%s %s channel(s)" % t for t in zip(counts, ch_types)
+                "%s '%s' channel(s)" % t for t in zip(counts, ch_types)
             )
-            msg = "Only one datatype can be fitted, but 'picks' results in "
-            msg += channels_msg
+            msg = (
+                "Only one datatype can be selected for fitting, but 'picks' "
+                f"results in {channels_msg}."
+            )
             raise ValueError(msg)
