### 2022-09-02 b083a81
test static method in test_base.py
