### 2022-09-02 4b3a9a7
rename  test_predict_default_picks to  test_predict_default since it tests also for defaults of other arguments
