### 2022-09-02 182b703
add a _check_unfitted to make sure the instance is not fitted before .fit
