+++ pycrostates/cluster/tests/test_kmeans.py
@@ -148,6 +148,7 @@ def test_ModKMeans():
     )
 
     # Test fit on ChData
+    ModK1.fitted = False
     ModK1.fit(ch_data, n_jobs=1)
     _check_fitted(ModK1)
     assert ModK1._cluster_centers_.shape == (
@@ -900,72 +901,61 @@ def test_picks_fit_predict(caplog):
         ModK_.fit(raw, picks="data")  # fails -> eeg + grad + mag
 
     # test valid fit
+    ModK_.fit(raw, picks="mag")  # works
+    ModK_.fitted = False
     ModK_.fit(raw, picks="eeg")  # works
     ModK_.fitted = False
-    ModK_.fit(raw, picks="mag")  # works
+
+    # create mock raw for fitting
     info_ = create_info(
         ["Fp1", "Fp2", "CP1", "CP2"], sfreq=1024, ch_types="eeg"
     )
     info_.set_montage("standard_1020")
     data = np.random.randn(4, 1024 * 10)
 
-    # ---- Ignore bad channel Fp2 during fitting ----
-    # Fit
+    # Ignore bad channel Fp2 during fitting
     info = info_.copy()
     info["bads"] = ["Fp2"]
     raw = RawArray(data, info)
 
-    ModK_ = ModK.copy()
+    caplog.clear()
     ModK_.fit(raw, picks="eeg")
     assert ModK_.info["ch_names"] == ["Fp1", "CP1", "CP2"]
-    assert "Channel Fp2 is set as bad and ignored" in caplog.text
+    assert "The channel Fp2 is set as bad and ignored" in caplog.text
     caplog.clear()
 
-    # Predict
-    # overall, same channels in the instance used for prediction
+    # predict with the same channels in the instance used for prediction
     info = info_.copy()
     raw_predict = RawArray(data, info)
-    ModK_.predict(
-        raw_predict, picks="eeg"
-    )  # -> works, with warning that Fp2 is missing from ModK.info
-    assert "Picked channel Fp2 will not be used for prediction" in caplog.text
     caplog.clear()
-
-    ModK_.predict(
-        raw_predict, picks=["Fp1", "CP1", "CP2"]
-    )  # -> works, same channels as ModK.info
+    ModK_.predict(raw_predict, picks="eeg")  # -> warning for selected Fp2
+    assert "Fp2 which was not used during fitting" in caplog.text
     caplog.clear()
-
-    raw_predict.info["bads"] = ["Fp2"]
-    ModK_.predict(
-        raw_predict, picks="eeg"
-    )  # -> works, same channels as ModK.info
-
-    raw_predict.info["bads"] = ["Fp1"]
-    with pytest.raises(
-        ValueError, match="Cannot predict because channel Fp1 is not in"
-    ):
-        ModK_.predict(raw_predict, picks="eeg")
-    # fails, because ModK.info includes Fp1
-    # which is bad in prediction instance
-    caplog.clear()
-
     ModK_.predict(raw_predict, picks=["Fp1", "CP1", "CP2"])
-    # works, with warning because you predict based on a bad channel Fp1
-    assert (
-        "Channel Fp1 is set as bad and will be used for prediction"
-        in caplog.text
-    )
+    assert "Fp2 which was not used during fitting" not in caplog.text
+    caplog.clear()
+    raw_predict.info["bads"] = ["Fp2"]
+    ModK_.predict(raw_predict, picks="eeg")
+    assert "Fp2 which was not used during fitting" not in caplog.text
+
+    # predict with a channel used for fitting that is now missing
+    # fails, because ModK.info includes Fp1 which is bad in prediction instance
+    raw_predict.info["bads"] = ["Fp1"]
[…]
