### 2022-09-02 e005879
handles non-deterministic set order in tests
