+++ tutorials/preprocessing/plot_resampling.py
@@ -35,8 +35,8 @@ raw.set_eeg_reference('average')
 #%%
 # We can now use the :func:`~pycrostates.preprocessing.resample` function to
 # draw n_resamples of n_samples for our recording where n_sample defines the
-# number of sample contained in each epoch and n_resamples defines the number of
-# epochs to draw.
+# number of sample contained in each epoch and n_resamples defines the number
+# of epochs to draw.
 
 from pycrostates.preprocessing import resample
 resamples = resample(raw, n_resamples=10, n_samples=150, random_state=40)
@@ -49,10 +49,13 @@ resamples
 
 resamples = resample(raw, n_resamples=10, coverage=0.5, random_state=40)
 resamples
-#%% or by setting n_samples and coverage:
+
+#%%
+# or by setting n_samples and coverage:
 
 resamples = resample(raw, n_samples=150, coverage=0.5, random_state=40)
 resamples
+
 #%%
 # Finally, we can also use this function to resample :func:`~mne.epochs.Epochs`
 
