### 2022-09-03 5d8a306
better config, green without plot_gallery
