+++ pyproject.toml
@@ -56,18 +56,19 @@ dependencies = [
 build = ['build', 'twine']
 docs = [
     'm2r2',
+    'memory-profiler',
     'nbsphinx',
     'numpydoc',
     'pandas',
     'pillow',
     'pydata-sphinx-theme',
     'pymatreader',
-    'recommonmark',
     'sphinx',
     'sphinxcontrib-bibtex',
     'sphinx-copybutton',
     'sphinx-design',
     'sphinx_gallery',
+    'sphinx-issues',
 ]
 style = [
     'black',
