### 2022-09-03 874d4be
use linkcode instead of viewcode and remove unused packages
