### 2022-09-03 e8d244a
set ploit_gallery back to True
