+++ docs/source/conf.py
@@ -288,7 +288,7 @@ sphinx_gallery_conf = {
     "examples_dirs": [str(Path(__file__).parent.parent.parent / "tutorials")],
     "gallery_dirs": ["generated/auto_tutorials"],
     "line_numbers": False,  # messes with style
-    "plot_gallery": False,
+    "plot_gallery": True,
     "reference_url": dict(pycrostates=None),  # documented lib uses None
     "remove_config_comments": True,
     "show_memory": sys.platform == "linux",
