### 2022-09-03 d7737da
re-enable plot_gallery and remove debug prints
