+++ docs/source/conf.py
@@ -5,12 +5,9 @@
 # https://www.sphinx-doc.org/en/master/usage/configuration.html
 
 
-import inspect
 import sys
 from datetime import date
-from importlib import import_module
 from pathlib import Path
-from typing import Dict, Optional
 
 from sphinx_gallery.sorting import ExplicitOrder, FileNameSortKey
 
@@ -44,8 +41,8 @@ extensions = [
     "sphinx.ext.autosectionlabel",
     "sphinx.ext.autosummary",
     "sphinx.ext.intersphinx",
-    "sphinx.ext.linkcode",
     "sphinx.ext.mathjax",
+    "sphinx.ext.viewcode",
     "numpydoc",
     "sphinxcontrib.bibtex",
     "sphinx_copybutton",
@@ -229,70 +226,6 @@ numpydoc_validation_exclude = {  # regex to ignore during docstring check
 # -- sphinxcontrib-bibtex ----------------------------------------------------
 bibtex_bibfiles = ["../references.bib"]
 
-# -- sphinx.ext.linkcode -----------------------------------------------------
-# https://www.sphinx-doc.org/en/master/usage/extensions/linkcode.html
-
-gh_urls = {
-    package: gh_url,
-    "mne": "https://github.com/mne-tools/mne-python",
-}
-
-
-def linkcode_resolve(domain: str, info: Dict[str, str]) -> Optional[str]:
-    """Determine the URL corresponding to a Python object.
-
-    Parameters
-    ----------
-    domain : str
-        One of 'py', 'c', 'cpp', 'javascript'.
-    info : dict
-        With keys "module" and "fullname".
-
-    Returns
-    -------
-    url : str | None
-        The code URL. If None, no link is added.
-    """
-    if domain != "py":
-        return None  # only document python objects
-
-    module = info["module"].split(".")[0]
-    if module not in gh_urls:
-        raise RuntimeError(
-            "This module is not configured in 'linkcode_resolve'. Please edit "
-            "the documentation 'conf.py' file."
-        )
-
-    # retrieve pyobject and file
-    try:
-        pyobject = import_module(info["module"])
-        for elt in info["fullname"].split("."):
-            pyobject = getattr(pyobject, elt)
-        while hasattr(pyobject, '__wrapped__'):
-            pyobject = pyobject.__wrapped__
-        fname = inspect.getsourcefile(pyobject).replace("\\", "/")
-    except Exception:
-        # Either the object could not be loaded or the file was not found.
-        # For instance, properties will raise.
-        return None
-
-    # retrieve start/stop lines
-    source, start_line = inspect.getsourcelines(pyobject)
-    lines = "L%d-L%d" % (start_line, start_line + len(source) - 1)
-
-    # create URL
-    if "dev" in release:
-        branch = "main"
-    else:
-        return None  # alternatively, link to a maint/version branch
-
-    while module in fname:
-        fname = fname.split(module)[-1]
-    assert fname[0] == "/"  # sanity-check, file separator
-    fname = fname[1:]
-    return f"{gh_urls[module]}/blob/{branch}/{package}/{fname}#{lines}"
-
-
 # -- sphinx-gallery ----------------------------------------------------------
 sphinx_gallery_conf = {
     "backreferences_dir": "generated/backreferences",
@@ -304,7 +237,7 @@ sphinx_gallery_conf = {
     },
     "gallery_dirs": ["generated/auto_tutorials"],
     "line_numbers": False,  # messes with style
-    "plot_gallery": True,
+    "plot_gallery": False,
     "reference_url": dict(pycrostates=None),  # documented lib uses None
     "remove_config_comments": True,
     "show_memory": sys.platform == "linux",
