### 2022-09-03 c669e1d
enable plot_gallery
