### 2022-09-03 fa57925
Update doc
Add Chdata to fit
Update Segmentation
