+++ pycrostates/cluster/kmeans.py
@@ -20,7 +20,7 @@ from ._base import _BaseCluster
 
 @fill_doc
 class ModKMeans(_BaseCluster):
-    r"""Modified K-Means clustering algorithms.
+    r"""Modified K-Means clustering algorithm.
 
     See :footcite:t:`391164` for additional information.
 
