+++ docs/source/api/cluster.rst
@@ -1,3 +1,5 @@
+.. _cluster:
+
 Cluster
 =======
 
