### 2022-09-03 71689c5
Update plot_group_level_with_individual_clusters.py
