### 2022-09-04 106df45
Update plot_group_level_with_individual_clusters.py
