### 2022-09-04 1f881e9
[DOC] Fix installation link in readme and improve documentation build (#58)
* fix link to install in readme

* improve doc build configuration

* update install page with a tab design

* move links to a separate file

* merge both links.inc

* fix some of the x-ref aliases

* add gh workflow to build the doc

* try with gh-action

* revert min sphinx version

* fix .gitignore

* fix doc references

* fix Array -> array

* move strictly postive mention

* remove gh doc-build

* fix title underline

* ignore doc validation for method that matches dict subclasses

* fix reference in clustering tu
[…]
