+++ docs/source/overview.rst
@@ -1,15 +1,0 @@
-Overview
-========
-
-An open-source Python package for analyzing EEG microstates developed to work
-with `MNE-python <https://mne.tools/stable/index.html>`_.
-
-.. image:: _static/img/FBMLAB_logo.png
-   :width: 200
-   :alt: FBMLAB logo
-   :target: https://www.unige.ch/medecine/neuf/en/researc/grecherche/christoph-michel
-
-.. image:: _static/img/UNIGE_logo.png
-   :width: 300
-   :alt: UNIGE logo
-   :target: https://www.unige.ch/en/university/presentation

[…]
