### 2022-09-04 8a19a93
Fix missing @fill_doc and delete unused %(verbose)s
