### 2022-09-04 1c98987
add version to doc conf.py
