+++ .readthedocs.yml
@@ -9,9 +9,6 @@ version: 2
 sphinx:
    configuration: docs/source/conf.py
 
-# Optionally build your docs in additional formats such as PDF
-formats: all
-
 # Optionally set the version of Python and requirements required to build your docs
 python:
    install:
