### 2022-09-04 5f59cb4
Fix default picks (#66)
* Fix default picks

* Update _docs.py

* Add _check_picks [WIP]

* Update fit picks

* Add pick to ChData

* Fix picks

* Add test_ChData_picks

* Add test_fit_picks

* Fix style

* Fix isort

* Fix style

* Fix style

* Update pycrostates/io/tests/test_ch_data.py

Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>

* Update pycrostates/cluster/_base.py

Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>

* Update pycrostates/cluster/_base.py

Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>

* Move docstring

* Upda
[…]
