+++ pycrostates/cluster/tests/test_kmeans.py
@@ -9,10 +9,10 @@ import numpy as np
 import pytest
 from matplotlib import pyplot as plt
 from matplotlib.figure import Figure
-from mne import Annotations, Epochs, make_fixed_length_events
+from mne import Annotations, Epochs, create_info, make_fixed_length_events
 from mne.channels import DigMontage
 from mne.datasets import testing
-from mne.io import read_raw_fif
+from mne.io import RawArray, read_raw_fif
 from mne.io.pick import _picks_to_idx
 
 from pycrostates import __version__
@@ -148,6 +148,7 @@ def test_ModKMeans():
     )
 
     # Test fit on ChData
+    ModK1.fitted = False
     ModK1.fit(ch_data, n_jobs=1)
     _check_fitted(ModK1)
     assert ModK1._cluster_centers_.shape == (
@@ -728,11 +729,9 @@ def test_fit_data_shapes():
     assert np.isclose(fitted_data_5_end, ModK_rej_5_end._fitted_data).all()
 
 
-def test_fit_with_bads(caplog):
-    """Test log messages emitted when fitting with bad channels."""
-    # 0 bads
-    raw_ = raw_eeg.copy()
-    raw_.info["bads"] = []
+def test_refit():
+    """Test that re-fit does not overwrite the current instance."""
+    raw = raw_meg.copy().pick_types(meg=True, eeg=True, eog=True)
     ModK_ = ModKMeans(
         n_clusters=n_clusters,
         n_init=10,
@@ -740,16 +739,18 @@ def test_fit_with_bads(caplog):
         tol=1e-4,
         random_state=1,
     )
-    ModK_.fit(raw_, n_jobs=1)
-    _check_fitted(ModK_)
-    _check_fitted_data_raw(ModK_._fitted_data, raw_, "eeg", None, None, "omit")
-    assert len(ModK_._info["bads"]) == 0
-    assert "set as bad" not in caplog.text
-    caplog.clear()
+    ModK_.fit(raw, picks="eeg")
+    eeg_ch_names = ModK_.info["ch_names"]
+    eeg_cluster_centers = ModK_.cluster_centers_
+    ModK_.fitted = False  # unfit
+    ModK_.fit(raw, picks="mag")
+    mag_ch_names = ModK_.info["ch_names"]
+    mag_cluster_centers = ModK_.cluster_centers_
+    assert eeg_ch_names != mag_ch_names
+    assert eeg_cluster_centers.shape != mag_cluster_centers.shape
 
-    # 1 bads
-    raw_ = raw_eeg.copy()
-    raw_.info["bads"] = [raw_.ch_names[0]]
+    # invalid
+    raw = raw_meg.copy().pick_types(meg=True, eeg=True, eog=True)
     ModK_ = ModKMeans(
         n_clusters=n_clusters,
         n_init=10,
@@ -757,49 +758,16 @@ def test_fit_with_bads(caplog):
         tol=1e-4,
         random_state=1,
     )
-    ModK_.fit(raw_, n_jobs=1)
-    _check_fitted(ModK_)
-    _check_fitted_data_raw(ModK_._fitted_data, raw_, "eeg", None, None, "omit")
-    assert len(ModK_._info["bads"]) == 1
-    assert "Channel EEG 001 is set as bad and ignored." in caplog.text
-    caplog.clear()
-
-    # more than 1 bads
-    raw_ = raw_eeg.copy()
-    raw_.info["bads"] = [raw_.ch_names[0], raw_.ch_names[1]]
-    ModK_ = ModKMeans(
-        n_clusters=n_clusters,
-        n_init=10,
-        max_iter=100,
-        tol=1e-4,
[…]

[…]
