### 2022-09-05 4e057f0
Add _check_picks_uniqueness for extract_gfp_peaks
