+++ pycrostates/utils/_checks.py
@@ -8,6 +8,7 @@ from pathlib import Path
 
 import numpy as np
 from matplotlib.axes import Axes
+from mne import pick_info
 from mne.utils import check_random_state
 
 
@@ -279,3 +280,18 @@ def _check_tmin_tmax(inst, tmin, tmax):
             f"{inst.times[-1]}s instance."
         )
     return tmin, tmax
+
+
+def _check_picks_uniqueness(info, picks):
+    info = pick_info(info, picks, copy=True)
+    if len(info.get_channel_types(unique=True)) != 1:
+        ch_types = info.get_channel_types(unique=False)
+        ch_types, counts = np.unique(ch_types, return_counts=True)
+        channels_msg = ", ".join(
+            "%s '%s' channel(s)" % t for t in zip(counts, ch_types)
+        )
+        msg = (
+            "Only one datatype can be selected for fitting, but 'picks' "
+            f"results in {channels_msg}."
+        )
+        raise ValueError(msg)
