### 2022-09-05 cd74290
Update pycrostates/utils/_checks.py
Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>
