### 2022-09-05 7b2bca6
Fix picks docstring
