### 2022-09-05 b2ef11b
Add return_all parameter
