### 2022-09-06 e041b7b
Update pycrostates/preprocessing/extract_gfp_peaks.py
Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>
