### 2022-09-06 a0739af
rework data picking to reduce memory footprint
