### 2022-09-06 9ae9f27
fix shapes and keyword argument
