### 2022-09-06 88cbc0e
don't use squeeze or it will also squeeze if there is only one channel
