### 2022-09-06 1fd7e19
retrieve information from dictionary more safely
