### 2022-09-06 2c5172e
add comparison of projs and coordinate transforms to comparison of Info which should fail the I/O test
