### 2022-09-07 4611f5a
remove part about HPI results, let's not keep them in a ChInfo
