### 2022-09-07 ea6f84a
Add _check_picks_uniqueness for extract_gfp_peaks (#73)
* Add _check_picks_uniqueness for extract_gfp_peaks

* Fix style

* Update pycrostates/utils/_checks.py

Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>

* Fix picks docstring

* Add return_all parameter

* Fix style

* Update pycrostates/preprocessing/extract_gfp_peaks.py

Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>

* Fix style

* simpler

* rework data picking to reduce memory footprint

* simpler

* fix shapes and keyword argument

* fix tests

* don't use squeeze or it will also squeeze if 
[…]
