### 2022-09-07 ca39525
correctly initialize "dev_head_t"
