### 2022-09-09 374f234
compare projections if MNE is above 1.2
