+++ pycrostates/io/tests/test_meas_info.py
@@ -1,6 +1,7 @@
 """Test meas_info.py"""
 
 from collections import OrderedDict
+from copy import deepcopy
 
 import numpy as np
 import pytest
@@ -9,6 +10,7 @@ from mne.channels import DigMontage
 from mne.datasets import testing
 from mne.io import read_raw_fif
 from mne.io.constants import FIFF
+from mne.utils import check_version
 from mne.transforms import Transform
 
 from pycrostates.io import ChInfo
@@ -103,7 +105,6 @@ def test_create_from_info():
 
     # test with a file with projs and coordinate transformation
     chinfo = ChInfo(info=raw.info)
-    assert chinfo["projs"] == raw.info["projs"]
     assert chinfo["ctf_head_t"] is None
     assert chinfo["dev_ctf_t"] is None
     assert chinfo["dev_head_t"] == raw.info["dev_head_t"]
@@ -388,3 +389,14 @@ def test_comparison(caplog):
 
     # with different object
     assert chinfo1 != 101
+
+    # with projs
+    info1 = ChInfo(raw.info)
+    info2 = ChInfo(raw.info)
+    assert chinfo1 == chinfo2
+    chinfo1["projs"][0] = deepcopy(chinfo1["projs"][1])
+
+    if check_version("mne", "1.2"):
+        assert chinfo1 != chinfo2
+    else:
+        assert chinfo1 == chinfo2
