### 2022-09-09 07f2dbd
Use invert=True instead of custom code in _annotations_starts_stops
