### 2022-09-09 5f38b32
Add CHANGELOG (#74)
* Add Dev section to documentation with CHANGELOG

* Update latest.inc

* Update docs/source/dev/changes/latest.inc

Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>

* Update docs/source/dev/changes/latest.inc

Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>

* Update docs/source/dev/changes/latest.inc

Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>

* Update docs/source/dev/changes/latest.inc

Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>

Co-authored-by: Victor Ferat <Victor.Ferat@unige.ch>
[…]
