+++ docs/source/dev/index.rst
@@ -1,0 +1,13 @@
+Dev
+===
+
+Pycrostates is maintained by `Victor Férat`_ and `Mathieu Scheltienne`_.
+
+.. include:: changes/names.inc
+
+.. toctree::
+   :hidden:
+
+   whats_new.rst
+
+
