+++ docs/references.bib
@@ -60,3 +60,17 @@
 	year = {2019},
 	pages = {180308},
 }
+
+# GFP
+@article{KOENIG20161104,
+title = {Inappropriate assumptions about EEG state changes and their impact on the quantification of EEG state dynamics},
+journal = {NeuroImage},
+volume = {125},
+pages = {1104-1106},
+year = {2016},
+issn = {1053-8119},
+doi = {https://doi.org/10.1016/j.neuroimage.2015.06.035},
+url = {https://www.sciencedirect.com/science/article/pii/S1053811915005388},
+author = {Thomas Koenig and Daniel Brandeis},
+keywords = {Microstates, Global field power, Topographic change}
+}
