+++ docs/source/microstates/glossary.rst
@@ -1,0 +1,13 @@
+Glossary
+========
+
+.. glossary::
+    :sorted:
+
+    GFP
+        Global Field Power. For EEG data, it is computed as the standard deviation of the sensors at a given timepoint.
+        Local maxima of the global field power (GFP) are known to represent the portions of EEG data with highest signal-to-noise ratio :cite:`KOENIG20161104`
+    cluster centers
+        Arithmetic mean of all the points belonging to the cluster.
+        In the context microstate analysis, it corresponds to the microstates topographies.
+    
