+++ docs/source/microstates/index.rst
@@ -1,0 +1,11 @@
+.. include:: ../links.inc
+
+Microstates
+===========
+
+.. toctree::
+   :hidden:
+
+   microstates.rst
+   glossary.rst
+   bibliography.rst
