+++ docs/references.bib
@@ -74,3 +74,13 @@ url = {https://www.sciencedirect.com/science/article/pii/S1053811915005388},
 author = {Thomas Koenig and Daniel Brandeis},
 keywords = {Microstates, Global field power, Topographic change}
 }
+
+@article{Marqui1995,
+  author={Pascual-Marqui, R.D. and Michel, C.M. and Lehmann, D.},
+  journal={IEEE Transactions on Biomedical Engineering}, 
+  title={Segmentation of brain electrical activity into microstates: model estimation and validation}, 
+  year={1995},
+  volume={42},
+  number={7},
+  pages={658-665},
+  doi={10.1109/10.391164}}
