### 2022-09-13 10fffd6
[WIP] improve metrics tutorial
