+++ tutorials/metrics/plot_metrics.py
@@ -1,16 +1,25 @@
 """
-Choosing the number of cluster centers k.
-=========================================
+Evaluating clustering results.
+==============================
 
-In this tutorial, we will see how to get indications on the best
-number of clusters centers k to use.
-It is important to note that there is no single solution to this choice.
-It will always be a trade off between the quality of the clustering,
-the possibility to analyze the microstates in the light of the literature
-and the variance expressed by the decomposition.
+Since microstate analysis is based on an empirical model,
+it is impossible to know the ground truth label of each sample.
+It is therefore necessary to rely on reliable metrics in order to evaluate
+the relevance of the parameters used in our model, and more particularly
+the number of clusters center.
 """
 
-#%%
+# %%
+# Pycrostates implements several metrics to evaluate the quality of
+# clustering solutions without knowing the ground truth:
+# - The silhouette score :py:func:`pycrostates.metrics.silhouette_score`(higher the better)
+# - The Calinski Harabasz score :py:func:`pycrostates.metrics.calinski_harabasz_score`(higher the better)
+# - The Dunn score :py:func:`pycrostates.metrics.dunn_score`(higher the better)
+# - The Davies Bouldin score :py:func:`pycrostates.metrics.davies_bouldin_score`(lower the better)
+#
+# Those metrics can directly be applied to a fitted clustering algortihm
+# such as the :py:class:`~pycrostates.cluster.ModKmeans`
+#
 # .. note::
 #
 #     The lemon datasets is composed of EEGLAB files. To use the MNE reader
@@ -23,6 +32,7 @@ and the variance expressed by the decomposition.
 #
 #     Note that an environment created via the MNE installers includes
 #     ``pymatreader`` by default.
+
 
 import matplotlib.pyplot as plt
 from mne.io import read_raw_eeglab
@@ -37,17 +47,17 @@ raw.crop(0, 30)
 raw.pick("eeg")
 raw.set_eeg_reference("average")
 
-#%%
-# Pycrostates implements several metrics to evaluate the quality of
-# clustering solutions without knowing the ground truth:
-# - The silhouette score (higher the better)
-# - The Calinski Harabasz score (higher the better)
-# - The Dunn score (higher the better)
-# - The Davies Bouldin score (lower the better)
-#
-# We can apply these metrics on clustering solution
-# computed with different values of n_clusters and
-# analyse which give the best clustering solution.
+# %%
+# In this exemple , we will use a single subject EEG recording in order to
+# study more give more explanation about each of this metrics
+# before showing an application case to determine the optimal number of cluster
+# (i.e microstate topographies) `n_clusters` while performing clustering with
+# the :py:class:`~pycrostates.cluster.ModKmeans` algorithm. 
+# For computation cost reasons, we start by computing
+# each of the score on the clustering reuslts of
+# a :py:class:`~pycrostates.cluster.ModKmeans` fitted for 
+# different values of `n_clusters`
+
 from pycrostates.metrics import (
     silhouette_score,
     calinski_harabasz_score,
@@ -74,7 +84,14 @@ for k in cluster_numbers:
     davies_bouldin_scores.append(davies_bouldin_score(ModK))
 
 #%%
-# Silhouette score
+# The silhouette score
+# --------------------
+# Silhouette analysis compromises two scores:
+# - The intra cluster distance
+# - The between cluster distance
+# Overall, it summurizes how well clusters are dense and well separated.
+# The silhouette score is bounded between ``-1`` for poor clustering
+# to ``1`` for well separated clusters.
 
 plt.figure()
 plt.scatter(cluster_numbers, silhouette_scores)
@@ -82,16 +99,37 @@ plt.xlabel('n_clusters')
[…]
