+++ docs/references.bib
@@ -75,6 +75,7 @@ author = {Thomas Koenig and Daniel Brandeis},
 keywords = {Microstates, Global field power, Topographic change}
 }
 
+# Segmentation
 @article{Marqui1995,
   author={Pascual-Marqui, R.D. and Michel, C.M. and Lehmann, D.},
   journal={IEEE Transactions on Biomedical Engineering}, 
@@ -84,3 +85,18 @@ keywords = {Microstates, Global field power, Topographic change}
   number={7},
   pages={658-665},
   doi={10.1109/10.391164}}
+
+# Metrics
+@article{ROUSSEEUW198753,
+title = {Silhouettes: A graphical aid to the interpretation and validation of cluster analysis},
+journal = {Journal of Computational and Applied Mathematics},
+volume = {20},
+pages = {53-65},
+year = {1987},
+issn = {0377-0427},
+doi = {https://doi.org/10.1016/0377-0427(87)90125-7},
+url = {https://www.sciencedirect.com/science/article/pii/0377042787901257},
+author = {Peter J. Rousseeuw},
+keywords = {Graphical display, cluster analysis, clustering validity, classification},
+abstract = {A new graphical display is proposed for partitioning techniques. Each cluster is represented by a so-called silhouette, which is based on the comparison of its tightness and separation. This silhouette shows which objects lie well within their cluster, and which ones are merely somewhere in between clusters. The entire clustering is displayed by combining the silhouettes into a single plot, allowing an appreciation of the relative quality of the clusters and an overview of the data configuration. The average silhouette width provides an evaluation of clustering validity, and might be used to select an ‘appropriate’ number of clusters.}
+}
