+++ pycrostates/metrics/silhouette.py
@@ -35,7 +35,7 @@ def silhouette_score(cluster):  # higher the better
 
     References
     ----------
-    .. [1] :cite:t: `ROUSSEEUW198753`
+    .. [1] :cite:cts:`Silhouettes`
     """
     _check_type(cluster, (_BaseCluster,), item_name="cluster")
     cluster._check_fit()
