+++ docs/references.bib
@@ -38,7 +38,7 @@
   keywords = {EEG microstates, Resting state networks, Consciousness, Psychiatric disease, State-dependent information processing, Metastability},
 }
 % Modified K-Means reference
-@article{391164,
+@article{Marqui1995,
   author={Pascual-Marqui, R.D. and Michel, C.M. and Lehmann, D.},
   journal={IEEE Transactions on Biomedical Engineering},
   title={Segmentation of brain electrical activity into microstates: model estimation and validation},
@@ -75,19 +75,8 @@ author = {Thomas Koenig and Daniel Brandeis},
 keywords = {Microstates, Global field power, Topographic change}
 }
 
-# Segmentation
-@article{Marqui1995,
-  author={Pascual-Marqui, R.D. and Michel, C.M. and Lehmann, D.},
-  journal={IEEE Transactions on Biomedical Engineering}, 
-  title={Segmentation of brain electrical activity into microstates: model estimation and validation}, 
-  year={1995},
-  volume={42},
-  number={7},
-  pages={658-665},
-  doi={10.1109/10.391164}}
-
 # Metrics
-@article{ROUSSEEUW198753,
+@article{Silhouettes,
 title = {Silhouettes: A graphical aid to the interpretation and validation of cluster analysis},
 journal = {Journal of Computational and Applied Mathematics},
 volume = {20},
@@ -100,3 +89,41 @@ author = {Peter J. Rousseeuw},
 keywords = {Graphical display, cluster analysis, clustering validity, classification},
 abstract = {A new graphical display is proposed for partitioning techniques. Each cluster is represented by a so-called silhouette, which is based on the comparison of its tightness and separation. This silhouette shows which objects lie well within their cluster, and which ones are merely somewhere in between clusters. The entire clustering is displayed by combining the silhouettes into a single plot, allowing an appreciation of the relative quality of the clusters and an overview of the data configuration. The average silhouette width provides an evaluation of clustering validity, and might be used to select an ‘appropriate’ number of clusters.}
 }
+
+@article{Calinski-Harabasz,
+author = {   T.   Caliński  and    J   Harabasz },
+title = {A dendrite method for cluster analysis},
+journal = {Communications in Statistics},
+volume = {3},
+number = {1},
+pages = {1-27},
+year  = {1974},
+publisher = {Taylor & Francis},
+doi = {10.1080/03610927408827101},
+URL = {https://www.tandfonline.com/doi/abs/10.1080/03610927408827101},
+eprint = {https://www.tandfonline.com/doi/pdf/10.1080/03610927408827101}
+}
+
+@article{Dunn,
+author = { J. C.   Dunn† },
+title = {Well-Separated Clusters and Optimal Fuzzy Partitions},
+journal = {Journal of Cybernetics},
+volume = {4},
+number = {1},
+pages = {95-104},
+year  = {1974},
+publisher = {Taylor & Francis},
+doi = {10.1080/01969727408546059},
+URL = {https://doi.org/10.1080/01969727408546059},
+eprint = {https://doi.org/10.1080/01969727408546059}
+}
+
+@article{Davies-Bouldin,
+author={Davies, David L. and Bouldin, Donald W.},
+journal={IEEE Transactions on Pattern Analysis and Machine Intelligence}, 
+title={A Cluster Separation Measure}, 
+year={1979},
+volume={PAMI-1},
+number={2},
+pages={224-227},
+doi={10.1109/TPAMI.1979.4766909}}
