+++ paper/paper.md
@@ -39,6 +39,8 @@ bibliography: paper.bib
 # Summary
 Microstate analysis of the electroencephalogram (EEG), introduced in 1972 by Lehman [@lehmann1971multichannel], is a spatiotemporal analysis technique that takes advantage of the full spatial resolution of EEG recordings. Formalized further by Pascual Marqui and colleagues [@pascual-marqui_segmentation_1995], microstate analysis studies the distribution of the surface EEG potential maps over time. It transforms the EEG recordings into sequences of successive states of variable duration, called EEG microstates.
 
+
+
 Pycrostates implements multiple modules that allow researchers to apply microstate analysis on their data:
 
 - the cluster module supports the different clustering algorithms that find the optimal topographic maps that sparsely represent the EEG. 
@@ -54,8 +56,8 @@ By design, Pycrostates natively supports all data types from MNE-Python [@gramfo
 
 # Statement of need
 
-Currently, several software and libraries are available to perform EEG microstates analysis: Cartool [@brunet_spatiotemporal_nodate], Thomas Koening's Matlab plugins (https://www.thomaskoenig.ch/index.php/software/microstates-in-eeglab), the Matlab microstates plugin [@poulsen_microstate_2018], or the python 2 library from Frederic von Wegner [@von_wegner_information-theoretical_2018]. In the last few years, the Python programming language ecosystem has expanded considerably, especially in scientific fields. The MNE-Python [@gramfort_meg_2013] library stands out for the analysis of human neurophysiological data. However, the current implementation of microstates analysis from Marijn van Vliet has a limited number of features and is not maintained on a regular basis. Pycrostates is based on the robust MNE-Python ecosystem and complements it with simple tools for developing integrated microstates analysis pipelines. 
-In addition to providing a simple and complete API that can replicate most of the analyses proposed in the literature [@MICHEL2018577], Pycrostates is built with a modular and scalable design, following modern development methods and standards to facilitate its maintenance and evolution over time. Finally, new users will benefit from the exhaustive documentation and tutorials describing the microstate analysis modules.
[…]
