+++ paper/paper.bib
@@ -143,3 +143,29 @@ abstract = {The present review discusses a well-established method for character
   year={1971},
   publisher={Elsevier}
 }
+
+@article{Makowski2021neurokit,
+    author = {Dominique Makowski and Tam Pham and Zen J. Lau and Jan C. Brammer and Fran{\c{c}}ois Lespinasse and Hung Pham and Christopher Schölzel and S. H. Annabel Chen},
+    title = {{NeuroKit}2: A Python toolbox for neurophysiological signal processing},
+    journal = {Behavior Research Methods},
+    volume = {53},
+    number = {4},
+    pages = {1689--1696},
+    publisher = {Springer Science and Business Media {LLC}},
+    doi = {10.3758/s13428-020-01516-y},
+    url = {https://doi.org/10.3758%2Fs13428-020-01516-y},
+    year = 2021,
+    month = {feb}
+}
+
+@software{patricia_milz_2016_48830,
+  author       = {Patricia Milz and
+                  Igor Gnatenko},
+  title        = {keypy: Advanced Functionalities},
+  month        = apr,
+  year         = 2016,
+  publisher    = {Zenodo},
+  version      = {v1.2.1-beta},
+  doi          = {10.5281/zenodo.48830},
+  url          = {https://doi.org/10.5281/zenodo.48830}
+}
