### 2022-09-20 0082773
Apply suggestions from code review
Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>
