+++ paper/paper.md
@@ -56,7 +56,7 @@ By design, Pycrostates natively supports all data types from MNE-Python [@gramfo
 
 # Statement of need
 
-Currently, several software and libraries are available to perform EEG microstates analysis: Cartool [@brunet_spatiotemporal_nodate], Thomas Koening's Matlab plugins (https://www.thomaskoenig.ch/index.php/software/microstates-in-eeglab), the Matlab microstates plugin [@poulsen_microstate_2018]. In the last few years, the Python programming language ecosystem has expanded considerably, especially in scientific fields. The MNE-Python [@gramfort_meg_2013] library stands out for the analysis of human neurophysiological data. Current python implementations of microstate analysis include the python 2 library from Frederic von Wegner [@von_wegner_information-theoretical_2018], Neurokit [@Makowski2021neurokit], Keypy [@patricia_milz_2016_48830](no longer maintained) and the MNE-microstates package (Marijn van Vliet). However, the current microstates implementation of microstates analysis has a limited number of features and/or not maintain in a regulara basis. For most of them, microstate segmentation are done at the subject level, and computation of microstates topographies and segmentations at the group are not covered. Pycrostates is based on the robust MNE-Python ecosystem and complements it with simple tools for developing integrated microstates analysis pipelines. It offering reusable classes, allowing group level analysis and advanced segmentation features such as smoothing and small segment rejection especially usefull for handling low signal to noise datapoints occuring during microstate polarity inversions. Therefore, it provides a simple and complete API that can replicate most of the analyses proposed in the literature [@MICHEL2018577] and by standard but non-python packages [@brunet_spatiotemporal_nodate] [@poulsen_microstate_2018]. Pycrostates is built with a modular and scalable design, following modern development methods and standards to facilitate its maintenance and evolution over time. All together, it offers a fertile ground for open source community to replicate current microstate methodology but also to make it evolve.
+Currently, several software and libraries are available to perform EEG microstates analysis: Cartool [@brunet_spatiotemporal_nodate], Thomas Koening's Matlab plugins (https://www.thomaskoenig.ch/index.php/software/microstates-in-eeglab), the Matlab microstates plugin [@poulsen_microstate_2018]. In the last few years, the Python programming language ecosystem has expanded considerably, especially in scientific fields. The MNE-Python [@gramfort_meg_2013] library stands out for the analysis of human neurophysiological data. Current python implementations of microstate analysis include the python 2 library from Frederic von Wegner [@von_wegner_information-theoretical_2018], Neurokit [@Makowski2021neurokit], Keypy [@patricia_milz_2016_48830](no longer maintained) and the MNE-microstates package (Marijn van Vliet). However, the current implementations of microstates analysis have a limited number of features and/or are not maintained on a regular basis. Most of them perform the microstate segmentation at the subject level. Computation of microstates topographies and segmentations at the group level are not covered. Pycrostates is based on the robust MNE-Python ecosystem and complements it with simple tools for developing integrated microstates analysis pipelines. It offers reusable classes, enabling group level analysis and advanced segmentation features such as smoothing and small segment rejection, especially usefull for handling low signal to noise datapoints occuring during microstate polarity inversions. Therefore, it provides a simple and complete API that can replicate most of the analyses proposed in the literature [@MICHEL2018577] and by standard but non-python packages [@brunet_spatiotemporal_nodate] [@poulsen_microstate_2018]. Pycrostates is built with a modular and scalable design, following modern development methods and standards to facilitate its maintenance and evolution over time. All together, it offers a fertile ground for open source community to replicate current microstate methodology but also to make it evolve.
 Finally, new users will benefit from the exhaustive documentation and tutorials describing the microstate analysis modules.
 
[…]
