+++ docs/references.bib
@@ -50,80 +50,74 @@
 }
 % Lemon dataset reference
 @article{babayan_mind-brain-body_2019,
-	title = {A mind-brain-body dataset of {MRI}, {EEG}, cognition, emotion, and peripheral physiology in young and old adults},
-	volume = {6},
-	issn = {2052-4463},
-	doi = {10.1038/sdata.2018.308},
-	number = {1},
-	journal = {Scientific Data},
-	author = {Babayan, Anahit and Erbey, Miray and Kumral, Deniz},
-	year = {2019},
-	pages = {180308},
+  title = {A mind-brain-body dataset of {MRI}, {EEG}, cognition, emotion, and peripheral physiology in young and old adults},
+  volume = {6},
+  issn = {2052-4463},
+  doi = {10.1038/sdata.2018.308},
+  number = {1},
+  journal = {Scientific Data},
+  author = {Babayan, Anahit and Erbey, Miray and Kumral, Deniz},
+  year = {2019},
+  pages = {180308},
 }
 
 # GFP
 @article{KOENIG20161104,
-title = {Inappropriate assumptions about EEG state changes and their impact on the quantification of EEG state dynamics},
-journal = {NeuroImage},
-volume = {125},
-pages = {1104-1106},
-year = {2016},
-issn = {1053-8119},
-doi = {https://doi.org/10.1016/j.neuroimage.2015.06.035},
-url = {https://www.sciencedirect.com/science/article/pii/S1053811915005388},
-author = {Thomas Koenig and Daniel Brandeis},
-keywords = {Microstates, Global field power, Topographic change}
+  title = {Inappropriate assumptions about EEG state changes and their impact on the quantification of EEG state dynamics},
+  journal = {NeuroImage},
+  volume = {125},
+  pages = {1104-1106},
+  year = {2016},
+  issn = {1053-8119},
+  doi = {10.1016/j.neuroimage.2015.06.035},
+  author = {Thomas Koenig and Daniel Brandeis},
+  keywords = {Microstates, Global field power, Topographic change}
 }
 
 # Metrics
 @article{Silhouettes,
-title = {Silhouettes: A graphical aid to the interpretation and validation of cluster analysis},
-journal = {Journal of Computational and Applied Mathematics},
-volume = {20},
-pages = {53-65},
-year = {1987},
-issn = {0377-0427},
-doi = {https://doi.org/10.1016/0377-0427(87)90125-7},
-url = {https://www.sciencedirect.com/science/article/pii/0377042787901257},
-author = {Peter J. Rousseeuw},
-keywords = {Graphical display, cluster analysis, clustering validity, classification},
-abstract = {A new graphical display is proposed for partitioning techniques. Each cluster is represented by a so-called silhouette, which is based on the comparison of its tightness and separation. This silhouette shows which objects lie well within their cluster, and which ones are merely somewhere in between clusters. The entire clustering is displayed by combining the silhouettes into a single plot, allowing an appreciation of the relative quality of the clusters and an overview of the data configuration. The average silhouette width provides an evaluation of clustering validity, and might be used to select an ‘appropriate’ number of clusters.}
+  title = {Silhouettes: A graphical aid to the interpretation and validation of cluster analysis},
+  journal = {Journal of Computational and Applied Mathematics},
+  volume = {20},
+  pages = {53-65},
+  year = {1987},
+  issn = {0377-0427},
+  doi = {10.1016/0377-0427(87)90125-7},
+  author = {Peter J. Rousseeuw},
[…]
