### 2022-09-26 e45a4b3
use similar formatting in paper.bib
