+++ paper/paper.bib
@@ -111,25 +111,25 @@ arXiv:1309.0238 [cs]},
 }
 
 @article{MICHEL2018577,
-title = {EEG microstates as a tool for studying the temporal dynamics of whole-brain neuronal networks: A review},
-journal = {NeuroImage},
-volume = {180},
-pages = {577-593},
-year = {2018},
-note = {Brain Connectivity Dynamics},
-issn = {1053-8119},
-doi = {10.1016/j.neuroimage.2017.11.062},
-url = {https://www.sciencedirect.com/science/article/pii/S105381191731008X},
-author = {Christoph M. Michel and Thomas Koenig},
-keywords = {EEG microstates, Resting state networks, Consciousness, Psychiatric disease, State-dependent information processing, Metastability},
-abstract = {The present review discusses a well-established method for characterizing resting-state activity of the human brain using multichannel electroencephalography (EEG).
-            This method involves the examination of electrical microstates in the brain, which are defined as successive short time periods during which the configuration of the scalp potential field remains semi-stable,
-			suggesting quasi-simultaneity of activity among the nodes of large-scale networks. A few prototypic microstates, which occur in a repetitive sequence across time, can be reliably identified across participants.
-			Researchers have proposed that these microstates represent the basic building blocks of the chain of spontaneous conscious mental processes, and that their occurrence and temporal dynamics determine the quality of mentation.
-			Several studies have further demonstrated that disturbances of mental processes associated with neurological and psychiatric conditions manifest as changes in the temporal dynamics of specific microstates. Combined EEG-fMRI
-			studies and EEG source imaging studies have indicated that EEG microstates are closely associated with resting-state networks as identified using fMRI. The scale-free properties of the time series of EEG microstates explain
-			why similar networks can be observed at such different time scales. The present review will provide an overview of these EEG microstates, available methods for analysis, the functional interpretations of findings regarding
-			these microstates, and their behavioral and clinical correlates.}
+  title = {EEG microstates as a tool for studying the temporal dynamics of whole-brain neuronal networks: A review},
+  journal = {NeuroImage},
+  volume = {180},
+  pages = {577-593},
+  year = {2018},
+  note = {Brain Connectivity Dynamics},
+  issn = {1053-8119},
+  doi = {10.1016/j.neuroimage.2017.11.062},
+  url = {https://www.sciencedirect.com/science/article/pii/S105381191731008X},
+  author = {Christoph M. Michel and Thomas Koenig},
+  keywords = {EEG microstates, Resting state networks, Consciousness, Psychiatric disease, State-dependent information processing, Metastability},
+  abstract = {The present review discusses a well-established method for characterizing resting-state activity of the human brain using multichannel electroencephalography (EEG).
+              This method involves the examination of electrical microstates in the brain, which are defined as successive short time periods during which the configuration of the scalp potential field remains semi-stable,
+		  	  suggesting quasi-simultaneity of activity among the nodes of large-scale networks. A few prototypic microstates, which occur in a repetitive sequence across time, can be reliably identified across participants.
+	  		  Researchers have proposed that these microstates represent the basic building blocks of the chain of spontaneous conscious mental processes, and that their occurrence and temporal dynamics determine the quality of mentation.
[…]
