+++ paper/paper.md
@@ -37,32 +37,85 @@ bibliography: paper.bib
 ---
 
 # Summary
-Microstate analysis of the electroencephalogram (EEG), introduced in 1972 by Lehman [@lehmann1971multichannel], is a spatiotemporal analysis technique that takes advantage of the full spatial resolution of EEG recordings. Formalized further by Pascual Marqui and colleagues [@pascual-marqui_segmentation_1995], microstate analysis studies the distribution of the surface EEG potential maps over time. It transforms the EEG recordings into sequences of successive states of variable duration, called EEG microstates.
 
+Microstate analysis of the electroencephalogram (EEG), introduced in 1972 by
+Lehman [@lehmann1971multichannel], is a spatiotemporal analysis technique that
+takes advantage of the full spatial resolution of EEG recordings. Formalized
+further by Pascual Marqui and colleagues [@pascual-marqui_segmentation_1995],
+microstate analysis studies the distribution of the surface EEG potential maps
+over time. It transforms the EEG recordings into sequences of successive states
+of variable duration, called EEG microstates.
 
+Pycrostates implements multiple modules that allow researchers to apply
+microstate analysis on their datasets:
 
-Pycrostates implements multiple modules that allow researchers to apply microstate analysis on their data:
-
-- the cluster module supports the different clustering algorithms that find the optimal topographic maps that sparsely represent the EEG. 
-- the segmentation module supports the study of microstates sequences and their summary measures.
-- the metrics module quantifies the quality of the fitted clustering algorithms.
+- the cluster module supports the different clustering algorithms that find the
+  optimal topographic maps that sparsely represent the EEG.
+- the segmentation module supports the study of microstates sequences and their
+  summary measures.
+- the metrics module quantifies the quality of the fitted clustering
+  algorithms.
 
 Additional modules help researchers to develop analysis pipelines:
 
-- the dataset module provides direct access to preprocessed data that can be used to test pipelines. As of writing, it supports the LEMON dataset [@babayan_mind-brain-body_2019] comprising preprocessed EEG recordings of 227 healthy participants.
-- the viz module provides visualization tools for performing microstate analyses.
+- the dataset module provides direct access to preprocessed data that can be
+  used to test pipelines. As of writing, it supports the LEMON dataset
+  [@babayan_mind-brain-body_2019] comprising preprocessed EEG recordings of 227
+  healthy participants.
+- the viz module provides visualization tools for performing microstate
+  analyses.
 
-By design, Pycrostates natively supports all data types from MNE-Python [@gramfort_meg_2013] and is not restricted to EEG data. It is designed to support future improvements and additions such as different clustering algorithms or new tools for sequence analysis such as Markov chains. Moreover, this work builds its API on top of the robust MNE-Python ecosystem, enabling a seamless integration of microstates analysis.
+By design, Pycrostates is not restricted to EEG data. Its API is build on top
+of the robust MNE-Python [@gramfort_meg_2013] ecosystem enabling a seamless
+integration of microstates analysis. It's modular design supports all data
+typesnfrom MNE-Python and future improvements and additions such as different
+clustering algorithms or new tools for sequences analysis such as Markov
+chains.
 
 # Statement of need
 
[…]
