### 2022-09-26 4055fbb
use default conf for bibtex
