+++ docs/source/microstates/index.rst
@@ -1,11 +1,8 @@
-.. include:: ../links.inc
-
-Microstates
-===========
-
 .. toctree::
    :hidden:
 
    microstates.rst
    glossary.rst
    bibliography.rst
+
+.. include:: microstates.rst
