+++ tutorials/metrics/plot_metrics.py
@@ -11,6 +11,7 @@ in our model, and more particularly the number of :term:`cluster centers`.
 # %%
 # Pycrostates implements several metrics to evaluate the quality of
 # clustering solutions without knowing the ground truth:
+#
 # * The Silhouette score :func:`~pycrostates.metrics.silhouette_score` (higher the better)
 # * The Calinski-Harabasz score :func:`~pycrostates.metrics.calinski_harabasz_score` (higher the better)
 # * The Dunn score :func:`~pycrostates.metrics.dunn_score` (higher the better)
@@ -32,12 +33,12 @@ in our model, and more particularly the number of :term:`cluster centers`.
 #     Note that an environment created via the MNE installers includes
 #     ``pymatreader`` by default.
 
-
 import matplotlib.pyplot as plt
 from mne.io import read_raw_eeglab
 
 from pycrostates.cluster import ModKMeans
 from pycrostates.datasets import lemon
+
 
 raw_fname = lemon.data_path(subject_id="010017", condition="EO")
 raw = read_raw_eeglab(raw_fname, preload=True)
@@ -52,9 +53,9 @@ raw.set_eeg_reference("average")
 # case to determine the optimal number of :term:`cluster centers`
 # ``n_clusters`` while fitting with the :class:`~pycrostates.cluster.ModKMeans`
 # algorithm.
-# For computation cost reasons, we start by computing each of the score on the
-# clustering results of a :class:`~pycrostates.cluster.ModKMeans` fitted for
-# different values of ``n_clusters``.
+# For this example, we start by computing each of the score on the clustering
+# results of a :class:`~pycrostates.cluster.ModKMeans` fitted for different
+# values of ``n_clusters``.
 
 from pycrostates.metrics import (
     silhouette_score,
@@ -63,94 +64,96 @@ from pycrostates.metrics import (
     davies_bouldin_score,
 )
 
-cluster_numbers = range(2,10)
-silhouette_scores = []
-calinski_harabasz_scores = []
-dunn_scores = []
-davies_bouldin_scores = []
-
-for k in cluster_numbers:
-    ModK = ModKMeans(n_clusters=k, random_state=42)
+cluster_numbers = range(2, 10)
+scores = dict(silhouette=[], calinski_harabasaz=[], dunn=[], davies_bouldin=[])
+for n_clusters in cluster_numbers:
+    # fit K-means algorithm with a set number of cluster centers
+    ModK = ModKMeans(n_clusters=n_clusters, random_state=42)
     ModK.fit(raw, n_jobs=2)
-    # silhouettee
-    silhouette_scores.append(silhouette_score(ModK))
-    # calinski and harabasz
-    calinski_harabasz_scores.append(calinski_harabasz_score(ModK))
-    # dunn
-    dunn_scores.append(dunn_score(ModK))
-    # davies bouldin
-    davies_bouldin_scores.append(davies_bouldin_score(ModK))
+
+    # compute scores
+    scores["silhouette"].append(silhouette_score(ModK))
+    scores["calinski_harabasaz"].append(calinski_harabasz_score(ModK))
+    scores["dunn"].append(dunn_score(ModK))
+    scores["davies_bouldin"].append(davies_bouldin_score(ModK))
 
 #%%
-# The silhouette score
+# The Silhouette score
 # --------------------
-# Silhouette analysis compromises two scores:
-# - The intra cluster distance
-# - The between cluster distance
-# Overall, it summurizes how well clusters are dense and well separated.
+# Silhouette analysis includes two measures:
+#
+# - intra-cluster distance
+# - inter-cluster distance
+#
+# Overall, it summarizes how well clusters are dense and well separated.
[…]
