### 2022-09-26 b324c19
improve note about the lemon dataset
