### 2022-09-26 f512541
format glossary
