+++ docs/source/microstates/glossary.rst
@@ -5,10 +5,12 @@ Glossary
     :sorted:
 
     GFP
-        Global Field Power. For EEG data, it is computed as the standard deviation of the sensors at a given timepoint.
-        Local maxima of the global field power (GFP) are known to represent the portions of EEG data with highest signal-to-noise ratio :cite:`KOENIG20161104`
-    
+        Global Field Power. For EEG data, it is computed as the standard
+        deviation of the sensors at a given timepoint. Local maxima of the
+        global field power (GFP) are known to represent the portions of EEG
+        data with highest signal-to-noise ratio :cite:`KOENIG20161104`
+
     cluster centers
-        Arithmetic mean of all the points belonging to the cluster.
-        In the context of microstate analysis, it corresponds to the microstates topographies.
-    
+        Arithmetic mean of all the points belonging to the cluster. In the
+        context of microstate analysis, it corresponds to the microstates
+        topographies.
