### 2022-09-27 3d2232b
clean-up comment style in references
