+++ docs/references.bib
@@ -3,7 +3,7 @@
 % If available, include a DOI (preferred) *or* a URL for a given reference, but
 # not both, as the DOI turns into a link which is redundant with the URL.
 
-% MNE-C reference
+% MNE-C
 @article{GramfortEtAl2014,
   title = {{{MNE}} Software for Processing {{MEG}} and {{EEG}} Data},
   author = {Gramfort, Alexandre and Luessi, Martin and Larson, Eric and Engemann, Denis A. and Strohmeier, Daniel and Brodbeck, Christian and Parkkonen, Lauri and H{\"a}m{\"a}l{\"a}inen, Matti S.},
@@ -13,7 +13,7 @@
   doi = {10.1016/j.neuroimage.2013.10.027},
   journal = {NeuroImage},
 }
-% MNE-Python reference
+% MNE-Python
 @article{GramfortEtAl2013a,
   title = {{{MEG}} and {{EEG}} Data Analysis with {{MNE}}-{{Python}}},
   author = {Gramfort, Alexandre and Luessi, Martin and Larson, Eric and Engemann, Denis A. and Strohmeier, Daniel and Brodbeck, Christian and Goj, Roman and Jas, Mainak and Brooks, Teon and Parkkonen, Lauri and H{\"a}m{\"a}l{\"a}inen, Matti S.},
@@ -24,7 +24,8 @@
   journal = {Frontiers in Neuroscience},
   number = {267}
 }
-% Michel 2018 reference
+
+% Michel 2018 - microstates review
 @article{MICHEL2018577,
   title = {EEG microstates as a tool for studying the temporal dynamics of whole-brain neuronal networks: A review},
   journal = {NeuroImage},
@@ -37,7 +38,7 @@
   author = {Christoph M. Michel and Thomas Koenig},
   keywords = {EEG microstates, Resting state networks, Consciousness, Psychiatric disease, State-dependent information processing, Metastability},
 }
-% Modified K-Means reference
+% Modified K-Means
 @article{Marqui1995,
   author={Pascual-Marqui, R.D. and Michel, C.M. and Lehmann, D.},
   journal={IEEE Transactions on Biomedical Engineering},
@@ -48,7 +49,8 @@
   pages={658-665},
   doi={10.1109/10.391164}
 }
-% Lemon dataset reference
+
+% Lemon datasets
 @article{babayan_mind-brain-body_2019,
   title = {A mind-brain-body dataset of {MRI}, {EEG}, cognition, emotion, and peripheral physiology in young and old adults},
   volume = {6},
@@ -61,7 +63,7 @@
   pages = {180308},
 }
 
-# GFP
+# Global Field Power
 @article{KOENIG20161104,
   title = {Inappropriate assumptions about EEG state changes and their impact on the quantification of EEG state dynamics},
   journal = {NeuroImage},
@@ -74,7 +76,7 @@
   keywords = {Microstates, Global field power, Topographic change}
 }
 
-# Metrics
+% Metrics
 @article{Silhouettes,
   title = {Silhouettes: A graphical aid to the interpretation and validation of cluster analysis},
   journal = {Journal of Computational and Applied Mathematics},
@@ -86,7 +88,6 @@
   author = {Peter J. Rousseeuw},
   keywords = {Graphical display, cluster analysis, clustering validity, classification},
 }
-
 @article{Calinski-Harabasz,
   author = {T. Caliński and J. Harabasz},
   title = {A dendrite method for cluster analysis},
@@ -98,7 +99,6 @@
   publisher = {Taylor & Francis},
   doi = {10.1080/03610927408827101},
 }
-
 @article{Dunn,
   author = {J. C. Dunn†},
[…]
