### 2022-09-27 9e3047f
more improvements to the tutorials
