+++ tutorials/group_level_analysis/plot_group_level_with_individual_clusters.py
@@ -7,6 +7,11 @@ by computing group-level topographies based on individual clusters.
 """
 
 #%%
+# .. Links
+#
+# .. _`mne installers`: https://mne.tools/stable/install/installers.html
+
+#%%
 # .. note::
 #
 #     The lemon datasets used in this tutorial is composed of EEGLAB files. To
@@ -17,7 +22,7 @@ by computing group-level topographies based on individual clusters.
 #     - ``pip install pymatreader``
 #     - ``conda install -c conda-forge pymatreader``
 #
-#     Note that an environment created via the MNE installers includes
+#     Note that an environment created via the `MNE installers`_. includes
 #     ``pymatreader`` by default.
 
 from mne.io import read_raw_eeglab
