### 2022-09-27 aca1b08
more improvements to the tutorials
