### 2022-09-27 152c97a
better gfp preprocessing tutorial
