+++ tutorials/metrics/plot_metrics.py
@@ -42,12 +42,18 @@ in our model, and more particularly the number of :term:`cluster centers`.
 #     Note that an environment created via the `MNE installers`_ includes
 #     ``pymatreader`` by default.
 
+# sphinx_gallery_thumbnail_number = 2
+
 import matplotlib.pyplot as plt
+import numpy as np
+from mne import set_log_level
 from mne.io import read_raw_eeglab
 
 from pycrostates.cluster import ModKMeans
 from pycrostates.datasets import lemon
 
+
+set_log_level("ERROR")  # reduce verbosity
 
 raw_fname = lemon.data_path(subject_id="010017", condition="EO")
 raw = read_raw_eeglab(raw_fname, preload=True)
@@ -55,7 +61,48 @@ raw.crop(0, 30)
 raw.pick("eeg")
 raw.set_eeg_reference("average")
 
+#%%
+# Silhouette score
+# ----------------
+# The Silhouette score\ :footcite:p:`Silhouettes` focuses on 2 metrics: the
+# intra-cluster distance and the inter-cluster distance. It summarizes how well
+# clusters are dense and well separated.
+#
+# The silhouette score is bounded between ``-1`` (low cluster separation)
+# to ``1`` (high cluster separation).
+
 # %%
+# Calinski-Harabasz score
+# -----------------------
+# The Calinski-Harabasz score\ :footcite:p:`Calinski-Harabasz` is the ratio of
+# the sum of inter-cluster dispersion and of intra-cluster dispersion for all
+# clusters (where the dispersion is defined as the sum of correlations
+# squared). As the silhouette score, it also summarizes how well clusters are
+# dense and well separated. Higher values indicates higher cluster density and
+# better separation.
+
+#%%
+# Dunn score
+# ----------
+#
+# The Dunn score\ :footcite:p:`Dunn` is defined as a ratio of the smallest
+# inter-cluster distance to the largest intra-cluster distance. Overall, it
+# summarizes how well clusters are farther apart and less dispersed. Higher
+# values indicates a better separation.
+
+#%%
+# Davies-Bouldin score
+# --------------------
+# The Davies-Bouldin score\ :footcite:p:`Davies-Bouldin` is defined as the
+# average similarity measure of each cluster with its most similar cluster,
+# where similarity is the ratio of intra-cluster distances to inter-cluster
+# distances. Overall, it summarizes how well clusters are farther apart and
+# less dispersed. Lower values indicates a better separation, ``0`` being the
+# lowest score possible.
+
+# %%
+# Compute scores
+# --------------
 # In this example, we will use a single subject EEG recording in order to give
 # more explanation about each of this metrics before showing an application
 # case to determine the optimal number of :term:`cluster centers`
@@ -73,95 +120,93 @@ from pycrostates.metrics import (
 )
 
 cluster_numbers = range(2, 9)
-scores = dict(silhouette=[], calinski_harabasaz=[], dunn=[], davies_bouldin=[])
-for n_clusters in cluster_numbers:
+scores = {
+    "Silhouette": np.zeros(len(cluster_numbers)),
+    "Calinski-Harabasaz": np.zeros(len(cluster_numbers)),
+    "Dunn": np.zeros(len(cluster_numbers)),
+    "Davies-Bouldin": np.zeros(len(cluster_numbers)),
+}
+for k, n_clusters in enumerate(cluster_numbers):
     # fit K-means algorithm with a set number of cluster centers
     ModK = ModKMeans(n_clusters=n_clusters, random_state=42)
-    ModK.fit(raw, n_jobs=2)
+    ModK.fit(raw, n_jobs=2, verbose="WARNING")
 
     # compute scores
-    scores["silhouette"].append(silhouette_score(ModK))
[…]
