### 2022-09-27 7d8957c
clean-up grouplevel tutorial
