### 2022-09-27 61e11a0
try different maps for metrics tutorial
