+++ docs/source/dev/whats_new.rst
@@ -1,9 +1,0 @@
-.. _whats_new:
-
-What's new
-==========
-
-.. currentmodule:: pycrostates
-
-.. include:: changes/latest.inc
-.. include:: changes/names.inc
