### 2022-09-27 95c201a
more improvements to the tutorials
