+++ tutorials/group_level_analysis/plot_group_level_with_individual_clusters.py
@@ -7,9 +7,7 @@ by computing group-level topographies based on individual clusters.
 """
 
 #%%
-# .. Links
-#
-# .. _`mne installers`: https://mne.tools/stable/install/installers.html
+# .. include:: ../../../../links.inc
 
 #%%
 # .. note::
