### 2022-09-27 f5d1dcb
better final sentence for group level tutorial
