+++ tutorials/group_level_analysis/plot_group_level_with_individual_gfp_peaks.py
@@ -73,8 +73,10 @@ ModK.rename_clusters(new_names=["A", "B", "C", "D", "F"])
 ModK.plot()
 
 #%%
-# We can now backfit the group level maps to each individual recording and
-# extract microstate parameters.
+# We can now use this fitted clustering algorithm to predict the segmentation
+# on each individual. This is also referred to as backfitting the group level
+# maps to each individual recording. Finally, we can extract microstate
+# parameters from the backfitted segmentations.
 
 import pandas as pd
 
