+++ tutorials/group_level_analysis/plot_group_level_with_individual_clusters.py
@@ -75,8 +75,10 @@ ModK.rename_clusters(new_names=["MS1", "MS2", "MS3", "MS4", "MS5"])
 ModK.plot()
 
 #%%
-# We can now backfit the group level maps to each individual recording and
-# extract microstate parameters.
+# We can now use this fitted clustering algorithm to predict the segmentation
+# on each individual. This is also referred to as backfitting the group level
+# maps to each individual recording. Finally, we can extract microstate
+# parameters from the backfitted segmentations.
 
 import pandas as pd
 
