### 2022-09-27 2b14e87
Add reference to glossary
Also change scores["Davies-Bouldin"]
