### 2022-09-27 2f52a41
Remove pandas in favor of matplotlib violinplots
