### 2022-09-28 6c405f8
split logos in separate folder
