+++ docs/source/index.rst
@@ -3,21 +3,21 @@
 Pycrostates
 ===========
 
-.. image:: ../_static/img/Pycrostates_logo_white.png
+.. image:: ../_static/logos/Pycrostates_logo_white.png
     :class: only-dark
 
-.. image:: ../_static/img/Pycrostates_logo_black.png
+.. image:: ../_static/logos/Pycrostates_logo_black.png
     :class: only-light
 
 An open-source Python package for analyzing EEG microstates developed to work
 with `MNE-python <mne stable_>`_.
 
-.. image:: ../_static/img/FBMLAB_logo.png
+.. image:: ../_static/logos/FBMLAB_logo.png
    :width: 200
    :alt: FBMLAB logo
    :target: https://www.unige.ch/medecine/neuf/en/researc/grecherche/christoph-michel
 
-.. image:: ../_static/img/UNIGE_logo.png
+.. image:: ../_static/logos/UNIGE_logo.png
    :width: 300
    :alt: UNIGE logo
    :target: https://www.unige.ch/en/university/presentation
