### 2022-09-28 931f9d7
add images of clusters distances to tutorial as well
