+++ docs/source/microstates/glossary.rst
@@ -24,22 +24,23 @@ Glossary
 
     inter-cluster distance
         Distance between two datapoints belonging to two different clusters.
-        Depending on the metric used, it can be computed in different ways, for example
-        (but no limited to) as the average distance between all datapoint belonging to
-        different clusters, the distance between two cluster centers, the minimal distance
-        between two datapoints belonging to different cluster.
+        Depending on the metric used, it can be computed in different ways, for
+        example (but no limited to) as the average distance between all
+        datapoint belonging to different clusters, the distance between two
+        :term:`cluster centers`, the minimal distance between two datapoints
+        belonging to different cluster.
 
         .. figure:: ../../_static/img/inter_cluster_distance.png
             :width: 600
             :alt: Inter-cluster distance illustration.
 
     intra-cluster distance
-        Distance between two datapoint belonging to same cluster. Depending on
-        the metric used, it can be computed in different ways, for example
+        Distance between two datapoint belonging to the same cluster. Depending
+        on the metric used, it can be computed in different ways, for example
         (but no limited to) as the average distance between all datapoint
-        belonging to same cluster, the average distance between all datapoint
-        at the cluster center or the maximal distance between two datapoints
-        belonging to the same cluster.
+        belonging to the same cluster, the average distance between all
+        datapoint at the cluster center or the maximal distance between two
+        datapoints belonging to the same cluster.
 
         .. figure:: ../../_static/img/intra_cluster_distance.png
             :width: 600
