+++ tutorials/clustering/plot_subject_level_with_resampling.py
@@ -23,6 +23,8 @@ the clustering results.
 #     Note that an environment created via the `MNE installers`_ includes
 #     ``pymatreader`` by default.
 
+import numpy as np
+from matplotlib import pyplot as plt
 from mne.io import read_raw_eeglab
 
 from pycrostates.cluster import ModKMeans
@@ -31,16 +33,16 @@ from pycrostates.datasets import lemon
 
 raw_fname = lemon.data_path(subject_id='010017', condition='EC')
 raw = read_raw_eeglab(raw_fname, preload=True)
-raw.crop(0, 30)
+raw.crop(0, 60)
 raw.pick('eeg')
 raw.set_eeg_reference('average')
 
 #%%
-# Resampling is a technique which consist of selecting a subset of
-# dataset several times. This method can be useful to study the
-# stability and reliability of clustering results.
-# In this example, we will split our data in ``n_resamples`` resamples each
-# containing ``n_samples`` randomly selected from the original recording.
+# Resampling is a technique which consist of selecting a subset from a
+# dataset several times. This method can be useful to study the stability and
+# reliability of clustering results. In this example, we will split our data in
+# ``n_resamples`` resamples each containing ``n_samples`` randomly selected
+# from the original recording.
 #
 # .. note::
 #
@@ -50,46 +52,60 @@ raw.set_eeg_reference('average')
 #
 #          gfp_peaks = pycrostates.preprocessing.extract_gfp_peaks(raw)
 #          resamples = resample(
-#              raw, n_resamples=10, n_samples=1000, random_state=42
+#              gfp_peaks, n_resamples=3, n_samples=1000, random_state=42
 #          )
 #
+
 from pycrostates.preprocessing import resample
 
-resamples = resample(
-    raw, n_resamples=10, n_samples=1000, random_state=42
-)
+resamples = resample(raw, n_resamples=3, n_samples=1000, random_state=42)
 
 #%%
-# We can compute the :term:`cluster centers` on each of the resample.
+# We can compute the :term:`cluster centers` on each of the resample and plot
+# the topographic maps fitted on a unique figure with the ``axes`` argument.
+
+f, ax = plt.subplots(nrows=len(resamples), ncols=5)
 
 resample_results = []
-for resamp in resamples:
+for k, resamp in enumerate(resamples):
+    # fit Modified K-means
     ModK = ModKMeans(n_clusters=5, random_state=42)
-    ModK.fit(resamp, n_jobs=2)
-    fig = ModK.plot(show=False)
-    fig.suptitle(f'GEV: {ModK.GEV_:.2f}', fontsize=20)
+    ModK.fit(resamp, n_jobs=2, verbose="WARNING")
     resample_results.append(ModK.cluster_centers_)
+    # plot the cluster centers
+    fig = ModK.plot(axes=ax[k, :])
+    plt.text(
+        0.5, 1.4, f"GEV: {ModK.GEV_:.2f}",
+        horizontalalignment='center',
+        verticalalignment='center',
+        transform=ax[k, 2].transAxes,
+        fontdict=dict(size=14),
+    )
+
+plt.subplots_adjust(top=0.9, hspace=0.5)
+plt.show()
 
 #%%
-# As we can see, each resampling clustering solution explains
-# about 70% of its GEV.
-# However, topographies can vary a lot from one solution to another.
+# Each resampling clustering solution explains about 70% of its Global
+# Explained Variance (:term:`GEV`). We can distinguish similar topographies
+# between fits, although with large variation, different signs and in a
+# different order.
 # The sparsity of results reveals how data sampling may influence
-# the results of microstates analysis. Sampling issue are inherent
-# to EEG recording (i.e recording have limited duration and are
-# done at a certain time of a day, in certain conditions..)
[…]
