### 2022-09-28 f3487b6
set light-dark mode in tutorial as well
