+++ tutorials/group_level_analysis/plot_group_level_with_individual_clusters.py
@@ -25,6 +25,7 @@ by computing group-level topographies based on individual clusters.
 
 # sphinx_gallery_thumbnail_number = 2
 
+import matplotlib.pyplot as plt
 import numpy as np
 from mne.io import read_raw_eeglab
 
@@ -95,15 +96,18 @@ for subject_id in subject_ids:
     ms_data.append(d)
 
 #%%
-# From this point on, we can start to visualize our results
-# and do your statistical analysis. For example we can
-# plot the :term:`GEV` of microstate class.
+# From this point on, we can visualize our results and do a statistical
+# analysis. For example we can plot the :term:`GEV` of each microstate class.
 
-import matplotlib.pyplot as plt
+data = [
+    [d['MS1_gev'], d['MS2_gev'], d['MS3_gev'], d['MS4_gev'], d['MS5_gev']]
+    for d in ms_data
+]
 
-data = np.array([[d['MS1_gev'], d['MS2_gev'], d['MS3_gev'], d['MS4_gev'], d['MS5_gev']] for d in ms_data])
-
-fig, ax = plt.subplots()
-ax.set_title('Global Explained Variance (%)')
-ax.violinplot(data)
-ax.set_xticklabels([""] + ModK.cluster_names)
+plt.violinplot(np.array(data))
+plt.title("Global Explained Variance (%)")
+plt.xticks(
+    ticks=range(1, len(ModK.cluster_names) + 1),
+    labels=ModK.cluster_names,
+)
+plt.show()
