### 2022-09-28 cc12f76
[Review] Improve documentation (#81)
* Improve docs

* Improve docs

* [WIP] improve metrics tutorial

* Improve docs

* Add metrcis citations

* Update paper

* Apply suggestions from code review

Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>

* minor fixes

* better references

* use similar formatting in paper.bib

* update paper phrasing

* try to fix biblio

* use default conf for bibtex

* improve main tutorial

* improve footcite

* Update plot_subject_level_with_resampling.py

* Update plot_subject_level_with_resampling.py

* fix x-ref

* more improvements

* fix 
[…]
