+++ docs/source/dev/changelog.rst
@@ -1,0 +1,9 @@
+.. _changelog:
+
+Changelog
+=========
+
+.. currentmodule:: pycrostates
+
+.. include:: changes/latest.inc
+.. include:: changes/names.inc
