+++ docs/source/microstates/index.rst
@@ -1,0 +1,8 @@
+.. toctree::
+   :hidden:
+
+   microstates.rst
+   glossary.rst
+   bibliography.rst
+
+.. include:: microstates.rst

[…]
