+++ tutorials/group_level_analysis/plot_group_level_with_individual_gfp_peaks.py
@@ -3,7 +3,8 @@ Group level analysis from individual GFP peaks
 ==============================================
 
 In this tutorial, we will learn how to conduct group level analysis
-by computing group level topographies based on individual :term:`GFP` peaks.
+by computing group level topographies based on individual
+:term:`Global Field Power` (:term:`GFP`) peaks.
 """
 
 #%%
