### 2022-09-28 d774a3b
add note for extract GFP peaks
