### 2022-09-28 vferat/pycrostates#64 @mscheltienne
Hello! Apologies for the long delay, we were lacking the bandwidth to work on the reviews.
We rework the documentation build in #58, notably including better x-ref and references to literature articles; and we improved the documentation and the tutorials in #81. See for instance the tutorial about the metrics module: https://pycrostates.readthedocs.io/en/latest/generated/auto_tutorials/metrics/plot_metrics.html#sphx-glr-generated-auto-tutorials-metrics-plot-metrics-py
In #83, I added a note about the use of the default values from `scipy.signal.find_peaks` with a x-ref to the `scipy` website.

Are those changes sufficient in your opinion?
