### 2022-09-28 vferat/pycrostates#51 @mscheltienne
Hello! Sorry for the long delay, we finally finalize modifications following the reviews and are working towards finalizing this submission.

About this issue:
- the phrasing of the note about `pymatreader` was improved and a linked to the MNE installers was added in #83.
- the jinja files are now correctly packaged as of 0.1.2
- the website build has been reworked and it now includes the updated `style.css` of MNE for better dark theme support.

Are those changes sufficient and can we close this issue?
