### 2022-09-28 vferat/pycrostates#60 @mscheltienne
Hello! Apologies for the long delay, but we finally had the time to address all the reviews.
The documentation build has been improved in #58, including notably better x-ref and literature referencing.
The documentation tutorials have been improved in #81, for instance, see:
- https://pycrostates.readthedocs.io/en/latest/generated/auto_tutorials/metrics/plot_metrics.html#sphx-glr-generated-auto-tutorials-metrics-plot-metrics-py
- https://pycrostates.readthedocs.io/en/latest/generated/auto_tutorials/preprocessing/plot_extract_gfp_peaks.html

which now includes more details and references to the literature.

The documentation is not cast into stone and will continue to evolve, but in the meantime, are those changes sufficient to resolve this issue?
