### 2022-09-28 897400a
Add note about find_peaks in the docstring for extract_gfp_peaks and improve glossary x-ref (#83)
* improve glossary

* better x-ref

* add note for extract GFP peaks
