+++ pycrostates/preprocessing/extract_gfp_peaks.py
@@ -31,10 +31,10 @@ def extract_gfp_peaks(
     reject_by_annotation: bool = True,
     verbose=None,
 ) -> CHData:
-    """:term:`GFP` peaks extraction.
+    """:term:`Global Field Power` (:term:`GFP`) peaks extraction.
 
-    Extract :term:`global field power` (GFP) peaks from :class:`~mne.Epochs` or
-    :class:`~mne.io.Raw`.
+    Extract :term:`Global Field Power` (:term:`GFP`) peaks from
+    :class:`~mne.Epochs` or :class:`~mne.io.Raw`.
 
     Parameters
     ----------
@@ -68,6 +68,13 @@ def extract_gfp_peaks(
     -------
     ch_data : ChData
         Samples at global field power peaks.
+
+    Notes
+    -----
+    The :term:`Global Field Power` (:term:`GFP`) peaks are extracted with
+    :func:`scipy.signal.find_peaks`. Only the ``distance`` argument is
+    filled with the value provided in ``min_peak_distance``. The other
+    arguments are set to their default values.
     """
     from ..io import ChData
 
