+++ docs/source/microstates/glossary.rst
@@ -5,17 +5,22 @@ Glossary
     :sorted:
 
     GFP
-        Global Field Power. For EEG data, it is computed as the standard
-        deviation of the sensors at a given timepoint. Local maxima of the
-        global field power (GFP) are known to represent the portions of EEG
-        data with highest signal-to-noise ratio (:cite:t:`KOENIG20161104`).
+    global field power
+        Global Field Power (GFP) is a measure of the (non-)uniformity of the
+        electromagnetic field at the sensors. It is typically calculated as the
+        standard deviation of the sensor values at each time point. Thus, it is
+        a one-dimensional time series capturing the spatial variability of the
+        signal across sensor locations. Local maxima of the global field power
+        (GFP) are known to represent the portions of EEG data with highest
+        signal-to-noise ratio (:cite:t:`KOENIG20161104`).
 
     GEV
+    global explained variance
         Global explained variance.
         Total explained variance expressed by a given state.
         It is computed as the sum of the explained variance
-        multiplied by the :term:`GFP` of each sample
-        assigned to a given state.
+        multiplied by the :term:`Global Field Power` (:term:`GFP`) of each
+        sample assigned to a given state.
 
     cluster centers
         Arithmetic mean of all the points belonging to the cluster. In the
@@ -35,7 +40,6 @@ Glossary
 
         .. image:: ../../_static/img/inter_cluster_distance_lm.png
             :class: only-light
-
 
     intra-cluster distance
         Distance between two datapoint belonging to the same cluster. Depending
