### 2022-09-30 vferat/pycrostates#65 @mscheltienne
Hello, thank you for the review, and apologies for the delay to address your comments. We were both lacking the bandwidth to work on this.

The paper has been updated in #81 and in #84.

> Pycrostates “is able to replicate most of the analyses proposed in the literature (Michel et al., 2018).”. Does this statement then imply that replicability is an issue facing the field that Pycrostates seeks to address?

This was not our intention. The goal was to mention that most pipelines proposed in this literature review are supported by `pycrostates`. The phrasing has been updated to reflect this in #84.

-------

We added mentions of other softwares (with the proper references) that focuses on or supports `microstates` analysis. The list was splitted between non-Python and Python softwares, with the latter:

> Current Python implementations of microstate analysis include the python 2 library from Frederic von Wegner [@von_wegner_information-theoretical_2018], Neurokit [@Makowski2021neurokit], Keypy [@patricia_milz_2016_48830] (no longer maintained) and the MNE-microstates package from Marijn van Vliet.

We did not include a benchmark between the softwares, as it will be very time-consuming to produce and requires to get familiar with each software first. This is however a long-term plan for the website documentation. As a starter, we will probably compared to `MNE-microstates` and to `NeuroKit` (c.f. #17).

We did however rework the paper to emphasize more the additional functionalities compared to most software (e.g. modularity supporting both subject and group level analysis, temporal smoothing, small segments rejection) and the integration within the MNE-ecosystem (similar structure with a familiar `info` attribute).

-------

Please let us know if those changes are sufficient.
