### 2022-09-30 #4564 @mscheltienne
Hello everyone! Apologies for the delay in the response and for the time we took to address your comments.
We addressed all the remaining reviews through several PR. To summarize, the issues still open are:
- @AJQuinn: https://github.com/vferat/pycrostates/issues/60
- @anshu-97: https://github.com/vferat/pycrostates/issues/65 and https://github.com/vferat/pycrostates/issues/64

We added comments specific to each issue directly on them, and hope the changes we made are sufficient to close them.
