+++ docs/source/dev/changelog.rst
@@ -6,4 +6,5 @@ Changelog
 .. currentmodule:: pycrostates
 
 .. include:: changes/latest.inc
+.. include:: changes/0.2.0.inc
 .. include:: changes/names.inc
