+++ docs/source/dev/changes/changelog.rst
@@ -1,0 +1,10 @@
+.. _changelog:
+
+Changelog
+=========
+
+.. toctree::
+    :titlesonly:
+
+    latest
+    0.2.0
