+++ docs/source/dev/changelog.rst
@@ -1,10 +1,0 @@
-.. _changelog:
-
-Changelog
-=========
-
-.. currentmodule:: pycrostates
-
-.. include:: changes/latest.inc
-.. include:: changes/0.2.0.inc
-.. include:: changes/names.inc
