+++ docs/source/dev/changes/0.2.0.rst
@@ -1,11 +1,13 @@
 .. _changes_0_2_0:
+.. include:: ./names.inc
 
-Version 0.2.0 (2022.09.30)
---------------------------
+Version 0.2.0 (2022.9.30)
+-------------------------
 
 Enhancements
 ~~~~~~~~~~~~
 - Improve tutorials and documentation. (:pr:`81` by `Victor Férat`_)
+
 
 API changes
 ~~~~~~~~~~~
