### 2022-10-03 512ffbc
[ENH] Rework Changelog (#86)
* Bump changelog

* clean-up changelog, remove unused sections

* Rework changelog

* Rename lastest to latest

Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>
