+++ docs/source/dev/development.rst
@@ -1,0 +1,15 @@
+.. include:: ../../links.inc
+.. include:: ./changes/names.inc
+
+Development
+===========
+
+Pycrostates is maintained by `Victor Férat`_ and `Mathieu Scheltienne`_.
+
+The project accepts contributions in the form of bug reports, fixes, feature
+additions, and documentation improvements (including typo corrections).
+The best way to start contributing is by `opening an issue <project issues_>`_
+on our GitHub page to discuss ideas for changes or enhancements, or to tell us
+about behavior that you think might be a bug.
+
+A complete changelog can be found :ref:`here <changelog>`.
