+++ docs/source/dev/changes/latest.rst
@@ -15,14 +15,16 @@
 
    Also add a corresponding entry for yourself in docs/dev/changes/names.inc
 
-.. _current:
+.. include:: ./names.inc
 
-Current (0.2.0.dev)
--------------------
+.. _latest:
+
+Current 0.3.0.dev
+-----------------
 
 Enhancements
 ~~~~~~~~~~~~
-- Improve tutorials and documentation. (:pr:`81` by `Victor Férat`_)
+
 
 Bugs
 ~~~~
@@ -30,6 +32,4 @@ Bugs
 
 API changes
 ~~~~~~~~~~~
-- Change default parameter ``min_peak_distance`` in :func:`pycrostates.preprocessing.extract_gfp_peaks` to 1. (:pr:`81` by `Victor Férat`_)
-- Change ``picks`` parameter in :meth:`pycrostates.cluster.ModKMeans.fit` (``pycrostates.cluster._BaseCluster.fit``) and :meth:`pycrostates.cluster.ModKMeans.predict` (``pycrostates.cluster._BaseCluster.predict``) (:pr:`66` by `Mathieu Scheltienne`_ and `Victor Férat`_)
-- Change ``picks`` parameter in :func:`pycrostates.preprocessing.extract_gfp_peaks` to force only one datatype in picks. Also add ``return_all`` parameter (:pr:`73` by `Mathieu Scheltienne`_ and `Victor Férat`_)
+
