### 2022-10-07 vferat/pycrostates#60 @AJQuinn
Hello @mscheltienne - no problem, apologies for the delay on my side as well.

Thank you for the changes, I think the additional support makes a big difference and I'm happy to sign off on the review in the other thread this afternoon. The updates to the 'evaluating cluster fits' tutorial really stand out - I think this level of accessability and detail reflects the quality of the code underneath.

I'm looking forward to seeing this continue to develop, congratulations on the package.
