### 2022-10-09 vferat/pycrostates#65 @anshu-97
Hi there - apologies for my delayed response as well! I agree that it'll indeed be time-consuming to make comparisons with all other available software and I note the emphasis in the added functions beyond that of alternative software is sufficient in my opinion. Looking forward to seeing more developments in your package!
