### 2022-10-10 ea6ac63
re-order the installation page and default to MNE-installers in the tab selection
