+++ docs/source/install.rst
@@ -3,6 +3,8 @@
 Install
 =======
 
+``pycrostates`` requires Python ``3.7`` or higher.
+
 Dependencies
 ------------
 
@@ -15,13 +17,22 @@ Dependencies
 * ``decorator``
 * ``jinja2``
 
-We require that you use Python ``3.7`` or higher.
-
 ``pycrostates`` works best with the latest stable release of MNE-Python. To
-ensure MNE-Python is up-to-date, see
+ensure MNE-Python is up-to-date, see the
 `MNE installation instructions <mne install_>`_.
 
+Methods
+-------
+
 .. tab-set::
+
+    .. tab-item:: MNE installers
+
+        As of MNE-Python 1.1, ``pycrostates`` is distributed in the
+        `MNE standalone installers <mne installers_>`_.
+
+        The installers create a conda environment with the entire MNE-ecosystem
+        setup, and more! This installation method is recommanded for beginners.
 
     .. tab-item:: Pypi
 
@@ -38,14 +49,6 @@ ensure MNE-Python is up-to-date, see
         .. code-block:: bash
 
             conda install -c conda-forge pycrostates
-
-    .. tab-item:: MNE installers
-
-        As of MNE-Python 1.1, ``pycrostates`` is distributed in the
-        `MNE standalone installers <mne installers_>`_.
-
-        The installers create a conda environment with the entire MNE-ecosystem
-        setup, and more!
 
     .. tab-item:: Snapshot of the current version
 
