+++ docs/source/api/index.rst
@@ -2,7 +2,6 @@
 
 .. _api_reference:
 
-=============
 API Reference
 =============
 
