### 2022-10-10 c31e471
improve module docstrings which appear in the API reference page
