+++ pycrostates/viz/__init__.py
@@ -1,4 +1,5 @@
-"""Viz module for Visualization routines."""
+"""Visualization routines that can be called directly or from methods of the
+main ``pycrostates`` classes."""
 
 from .cluster_centers import plot_cluster_centers
 from .segmentation import plot_epoch_segmentation, plot_raw_segmentation
