### 2022-10-12 #4564 @Kevin-Mattheus-Moerman
@vferat I have just read through your paper as well and all seems in order, and we will proceed to process this work for acceptance in JOSS. Please have a look at the below steps.

There are some minor editorial points remaining for your to consider:
- [ ] Find and replace all occurrences of `python` with `Python`, i.e. use a capital letter.
- [ ] Suggestion: replace `open-source` with `open source`.
- [ ] Suggestion: replace `state of the art` with `state of the art`.

Once you've considered the above. Please do the following:
- [ ] Call `@editorialbot generate pdf` to update the paper draft
- [ ] Archive a copy of the reviewed work on [ZENODO](https://zenodo.org/). We recommend that you do so for a newly minted/tagged release, however this is up to you. Some find these automated steps useful: https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content
- [ ] Once the archive is created please check if the title and the author names match those of the paper. You may need to manually edit your archive entry to change the title/author entries. You can also check if you can have it refer to the license properly (or leave it as Other(Open) since the archived content contains the license too). Next you may wish to add author ORCID information too. The most important things though are the title and author list.
- [ ] Once the above is complete please report back here with the archive DOI and which version tag we should use, i.e. are we still at 0.1.1 or have you created a new release.
