### 2022-10-12 #4564 @mscheltienne
Thanks @Kevin-Mattheus-Moerman! I've addressed your comments in https://github.com/vferat/pycrostates/pull/88
Could you please detail this one:

> Suggestion: replace state of the art with state of the art.

I've read it 3 times, and I can't see the difference 😉

The package is now in version `0.2.0`, which is also the version released as part of the `mne-installers` as of 1.2 (released today). It's very recent, so I don't think we need to cut a new release.
