### 2022-10-12 4f2199b
[JOSS] Editor review (#88)
* replace python with Python

* remove - in open-source

* state of the art with hyphens
