### 2022-10-13 #4564 @vferat
Hey @Kevin-Mattheus-Moerman! Thanks for the update

As Mathieu said, we have address the editorial comments in https://github.com/vferat/pycrostates/pull/88 which have been merged.

You can use the tag of the last release `0.2.0` as well as the corresponding DOI  [`10.5281/zenodo.7129852 `](https://doi.org/10.5281/zenodo.7129852) with the article. Since this release, we have not made any significant changes to the library (except addressing editorial comments and updating the changelog).

Let us know if you need anything else

Current Zenodo citation:
_Victor Férat, Mathieu Scheltienne, AJQuinn, & Lou. (2022). vferat/pycrostates: 0.2.0 (0.2.0). Zenodo. https://doi.org/10.5281/zenodo.7129852_
